% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_arc_layer.R
\name{add_arc_layer}
\alias{add_arc_layer}
\title{Add a \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/arc-layer.md}{ArcLayer} deck.gl layer to an \link{rdeck} map.}
\usage{
add_arc_layer(
  rdeck,
  id = NULL,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  get_source_position = NULL,
  get_target_position = NULL,
  get_source_color = NULL,
  get_target_color = NULL,
  get_width = NULL,
  get_height = NULL,
  get_tilt = NULL,
  width_units = "pixels",
  width_scale = 1,
  width_min_pixels = 0,
  width_max_pixels = 9007199254740991,
  ...
)
}
\arguments{
\item{rdeck}{\code{\link{rdeck}}
An \link{rdeck} map.}

\item{id}{\code{\link{character}}
The id of the layer. Layer ids must be unique per layer \code{type} for deck.gl
to properly distinguish between them.}

\item{data}{\code{\link{data.frame}} | \code{\link[sf:sf]{sf::sf}}}

\item{visible}{\code{\link{logical}}}

\item{pickable}{\code{\link{logical}}}

\item{opacity}{\code{\link{numeric}}}

\item{position_format}{\code{XY} | \code{XYZ}}

\item{color_format}{\code{RGB} | \code{RGBA}}

\item{auto_highlight}{\code{\link{logical}}}

\item{highlight_color}{\code{\link{integer}}}

\item{get_source_position}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{get_target_position}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{get_source_color}{accessor}

\item{get_target_color}{accessor}

\item{get_width}{accessor | \code{\link{numeric}}}

\item{get_height}{accessor | \code{\link{numeric}}}

\item{get_tilt}{accessor | \code{\link{numeric}}}

\item{width_units}{\code{pixels} | \code{meters}}

\item{width_scale}{\code{\link{numeric}}}

\item{width_min_pixels}{\code{\link{numeric}}}

\item{width_max_pixels}{\code{\link{numeric}}}

\item{...}{
  Arguments passed on to \code{\link[=arc_layer]{arc_layer}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\code{\link{rdeck}}
The \link{rdeck} map.
}
\description{
Add a \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/arc-layer.md}{ArcLayer} deck.gl layer to an \link{rdeck} map.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/arc-layer.md}
}
