% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bitmap_layer.R
\name{add_bitmap_layer}
\alias{add_bitmap_layer}
\title{Add BitmapLayer to an rdeck map.}
\usage{
add_bitmap_layer(
  rdeck,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  image = NULL,
  bounds = c(1, 0, 0, 1),
  desaturate = 0,
  transparent_color = c(0, 0, 0, 0),
  tint_color = c(255, 255, 255),
  ...
)
}
\arguments{
\item{rdeck}{\\verb{\{rdeck\}\\} an rdeck widget instance}

\item{data}{\code{{data.frame | sf}}}

\item{visible}{\code{{logical}}}

\item{pickable}{\code{{logical}}}

\item{opacity}{\code{{numeric}}}

\item{position_format}{\code{{"XY" | "XYZ"}}}

\item{color_format}{\code{{"RGB" | "RGBA"}}}

\item{auto_highlight}{\code{{logical}}}

\item{highlight_color}{\code{{integer}}}

\item{image}{\code{{"image" | "character"}}}

\item{bounds}{\code{{numeric}}}

\item{desaturate}{\code{{numeric}}}

\item{transparent_color}{\code{{integer}}}

\item{tint_color}{\code{{integer}}}

\item{...}{additional layer parameters to pass to deck.gl}
}
\value{
\\verb{\{rdeck\}\\}
}
\description{
Add BitmapLayer to an rdeck map.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/bitmap-layer.md}
}
