% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_contour_layer.R
\name{add_contour_layer}
\alias{add_contour_layer}
\title{Add ContourLayer to an rdeck map.}
\usage{
add_contour_layer(
  rdeck,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  cell_size = 1000,
  get_position = NULL,
  get_weight = NULL,
  gpu_aggregation = TRUE,
  aggregation = "SUM",
  contours = c(NULL),
  z_offset = 0.005,
  ...
)
}
\arguments{
\item{rdeck}{\\verb{\{rdeck\}\\} an rdeck widget instance}

\item{data}{\code{{data.frame | sf}}}

\item{visible}{\code{{logical}}}

\item{pickable}{\code{{logical}}}

\item{opacity}{\code{{numeric}}}

\item{position_format}{\code{{"XY" | "XYZ"}}}

\item{color_format}{\code{{"RGB" | "RGBA"}}}

\item{auto_highlight}{\code{{logical}}}

\item{highlight_color}{\code{{integer}}}

\item{cell_size}{\code{{numeric}}}

\item{get_position}{\code{{accessor | JS}}}

\item{get_weight}{\code{{accessor | JS}}}

\item{gpu_aggregation}{\code{{logical}}}

\item{aggregation}{\code{{"SUM" | "MEAN" | "MIN" | "MAX"}}}

\item{contours}{\code{{list}}}

\item{z_offset}{\code{{numeric}}}

\item{...}{additional layer parameters to pass to deck.gl}
}
\value{
\\verb{\{rdeck\}\\}
}
\description{
Add ContourLayer to an rdeck map.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/contour-layer.md}
}
