% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_grid_cell_layer.R
\name{add_grid_cell_layer}
\alias{add_grid_cell_layer}
\title{Add GridCellLayer to an rdeck map.}
\usage{
add_grid_cell_layer(
  rdeck,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  disk_resolution = 20,
  vertices = NULL,
  radius = 1000,
  angle = 0,
  offset = c(1, 1),
  coverage = 1,
  elevation_scale = 1,
  line_width_units = "meters",
  line_width_scale = 1,
  line_width_min_pixels = 0,
  line_width_max_pixels = 9007199254740991,
  extruded = TRUE,
  wireframe = FALSE,
  filled = TRUE,
  stroked = FALSE,
  get_position = NULL,
  get_fill_color = NULL,
  get_line_color = NULL,
  get_line_width = NULL,
  get_elevation = NULL,
  material = TRUE,
  cell_size = 1000,
  ...
)
}
\arguments{
\item{rdeck}{\\verb{\{rdeck\}\\} an rdeck widget instance}

\item{data}{\code{{data.frame | sf}}}

\item{visible}{\code{{logical}}}

\item{pickable}{\code{{logical}}}

\item{opacity}{\code{{numeric}}}

\item{position_format}{\code{{"XY" | "XYZ"}}}

\item{color_format}{\code{{"RGB" | "RGBA"}}}

\item{auto_highlight}{\code{{logical}}}

\item{highlight_color}{\code{{integer}}}

\item{disk_resolution}{\code{{numeric}}}

\item{vertices}{\code{{list}}}

\item{radius}{\code{{numeric}}}

\item{angle}{\code{{numeric}}}

\item{offset}{\code{{numeric}}}

\item{coverage}{\code{{numeric}}}

\item{elevation_scale}{\code{{numeric}}}

\item{line_width_units}{\code{{"pixels" | "meters"}}}

\item{line_width_scale}{\code{{numeric}}}

\item{line_width_min_pixels}{\code{{numeric}}}

\item{line_width_max_pixels}{\code{{numeric}}}

\item{extruded}{\code{{logical}}}

\item{wireframe}{\code{{logical}}}

\item{filled}{\code{{logical}}}

\item{stroked}{\code{{logical}}}

\item{get_position}{\code{{accessor | JS}}}

\item{get_fill_color}{\code{{accessor}}}

\item{get_line_color}{\code{{accessor}}}

\item{get_line_width}{\code{{accessor | numeric}}}

\item{get_elevation}{\code{{accessor | numeric}}}

\item{material}{\code{{logical}}}

\item{cell_size}{\code{{numeric}}}

\item{...}{additional layer parameters to pass to deck.gl}
}
\value{
\\verb{\{rdeck\}\\}
}
\description{
Add GridCellLayer to an rdeck map.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/grid-cell-layer.md}
}
