% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_layer.R
\name{add_grid_layer}
\alias{add_grid_layer}
\title{GridLayer}
\usage{
add_grid_layer(
  rdeck,
  id = "GridLayer",
  data = data.frame(),
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  color_domain = NULL,
  color_range = c("#ffffb2", "#fed976", "#feb24c", "#fd8d3c", "#f03b20", "#bd0026"),
  get_color_weight = 1,
  color_aggregation = "SUM",
  elevation_domain = NULL,
  elevation_range = c(0, 1000),
  get_elevation_weight = 1,
  elevation_aggregation = "SUM",
  elevation_scale = 1,
  cell_size = 1000,
  coverage = 1,
  get_position = position,
  extruded = FALSE,
  material = TRUE,
  get_color_value = NULL,
  lower_percentile = 0,
  upper_percentile = 100,
  color_scale_type = "quantize",
  get_elevation_value = NULL,
  elevation_lower_percentile = 0,
  elevation_upper_percentile = 100,
  elevation_scale_type = "linear",
  grid_aggregator = NULL,
  gpu_aggregation = FALSE,
  ...
)
}
\arguments{
\item{rdeck}{\code{rdeck}}

\item{id}{\code{character}
The id of the layer. Layer ids must be unique per layer \code{type} for deck.gl
to properly distinguish between them.}

\item{data}{\code{data.frame} | \code{sf::sf}
Deck.gl layer data, either a \link{data.frame} or \link[sf:sf]{sf::sf}.}

\item{visible}{\code{logical}
Whether the layer is visible.}

\item{pickable}{\code{logical}
Whether the layer responds to mouse pointer picking events.}

\item{opacity}{\code{numeric}
The opacity of the layer.}

\item{position_format}{\code{"XY"} | \code{"XYZ"}
Determines whether each coordinate has two (XY) or three (XYZ) numbers.}

\item{color_format}{\code{"RGB"} | \code{"RGBA"}
RGB will make the layer ignore the alpha channel of colours returned by
accessors. Opacity controlled by \code{opacity} is still applied.}

\item{auto_highlight}{\code{logical}
When \code{TRUE}, current object pointed by mouse pointer (when hovered over) is
highlighted with highlight_color. Requires \code{pickable} to be \code{TRUE}.}

\item{highlight_color}{\code{integer} vector of (R,G,B) or (R,G,B,A)
RGBA color to be used to render highlighted object. When 3 component (RGB)
array is specified, a default value of 255 is used for alpha.}

\item{...}{additional deck.gl layer parameters}
}
\description{
GridLayer
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.1.0/docs/layers/grid-layer.md}

Other add_layers: 
\code{\link{add_arc_layer}()},
\code{\link{add_bitmap_layer}()},
\code{\link{add_column_layer}()},
\code{\link{add_contour_layer}()},
\code{\link{add_cpu_grid_layer}()},
\code{\link{add_geojson_layer}()},
\code{\link{add_gpu_grid_layer}()},
\code{\link{add_great_circle_layer}()},
\code{\link{add_grid_cell_layer}()},
\code{\link{add_h3_cluster_layer}()},
\code{\link{add_h3_hexagon_layer}()},
\code{\link{add_heatmap_layer}()},
\code{\link{add_hexagon_layer}()},
\code{\link{add_icon_layer}()},
\code{\link{add_line_layer}()},
\code{\link{add_mvt_layer}()},
\code{\link{add_path_layer}()},
\code{\link{add_point_cloud_layer}()},
\code{\link{add_polygon_layer}()},
\code{\link{add_s2_layer}()},
\code{\link{add_scatterplot_layer}()},
\code{\link{add_scenegraph_layer}()},
\code{\link{add_screen_grid_layer}()},
\code{\link{add_simple_mesh_layer}()},
\code{\link{add_solid_polygon_layer}()},
\code{\link{add_terrain_layer}()},
\code{\link{add_text_layer}()},
\code{\link{add_tile3d_layer}()},
\code{\link{add_tile_layer}()},
\code{\link{add_trips_layer}()}
}
\concept{add_layers}
