% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hexagon_layer.R
\name{add_hexagon_layer}
\alias{add_hexagon_layer}
\title{Add HexagonLayer to an rdeck map.}
\usage{
add_hexagon_layer(
  rdeck,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  color_domain = NULL,
  color_range = list(c(255, 255, 178), c(254, 217, 118), c(254, 178, 76), c(253, 141,
    60), c(240, 59, 32), c(189, 0, 38)),
  get_color_value = NULL,
  get_color_weight = NULL,
  color_aggregation = "SUM",
  lower_percentile = 0,
  upper_percentile = 100,
  color_scale_type = "quantize",
  elevation_domain = NULL,
  elevation_range = c(0, 1000),
  get_elevation_value = NULL,
  get_elevation_weight = NULL,
  elevation_aggregation = "SUM",
  elevation_lower_percentile = 0,
  elevation_upper_percentile = 100,
  elevation_scale = 1,
  elevation_scale_type = "linear",
  radius = 1000,
  coverage = 1,
  extruded = FALSE,
  hexagon_aggregator = NULL,
  get_position = NULL,
  material = TRUE,
  ...
)
}
\arguments{
\item{rdeck}{\\verb{\{rdeck\}\\} an rdeck widget instance}

\item{data}{\code{{data.frame | sf}}}

\item{visible}{\code{{logical}}}

\item{pickable}{\code{{logical}}}

\item{opacity}{\code{{numeric}}}

\item{position_format}{\code{{"XY" | "XYZ"}}}

\item{color_format}{\code{{"RGB" | "RGBA"}}}

\item{auto_highlight}{\code{{logical}}}

\item{highlight_color}{\code{{integer}}}

\item{color_domain}{\code{{numeric}}}

\item{color_range}{\code{{list}}}

\item{get_color_value}{\code{{accessor}}}

\item{get_color_weight}{\code{{accessor | JS}}}

\item{color_aggregation}{\code{{"SUM" | "MEAN" | "MIN" | "MAX"}}}

\item{lower_percentile}{\code{{numeric}}}

\item{upper_percentile}{\code{{numeric}}}

\item{color_scale_type}{\code{{"quantize" | "linear" | "quantile" | "ordinal"}}}

\item{elevation_domain}{\code{{numeric}}}

\item{elevation_range}{\code{{numeric}}}

\item{get_elevation_value}{\code{{accessor}}}

\item{get_elevation_weight}{\code{{accessor | JS}}}

\item{elevation_aggregation}{\code{{"SUM" | "MEAN" | "MIN" | "MAX"}}}

\item{elevation_lower_percentile}{\code{{numeric}}}

\item{elevation_upper_percentile}{\code{{numeric}}}

\item{elevation_scale}{\code{{numeric}}}

\item{elevation_scale_type}{\code{{"quantize" | "linear" | "quantile" | "ordinal"}}}

\item{radius}{\code{{numeric}}}

\item{coverage}{\code{{numeric}}}

\item{extruded}{\code{{logical}}}

\item{hexagon_aggregator}{\code{{JS}}}

\item{get_position}{\code{{accessor | JS}}}

\item{material}{\code{{logical}}}

\item{...}{additional layer parameters to pass to deck.gl}
}
\value{
\\verb{\{rdeck\}\\}
}
\description{
Add HexagonLayer to an rdeck map.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/hexagon-layer.md}
}
