% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_icon_layer.R
\name{add_icon_layer}
\alias{add_icon_layer}
\title{Add IconLayer to an rdeck map.}
\usage{
add_icon_layer(
  rdeck,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  icon_atlas = NULL,
  icon_mapping = NULL,
  size_scale = 1,
  billboard = TRUE,
  size_units = "pixels",
  size_min_pixels = 0,
  size_max_pixels = 9007199254740991,
  alpha_cutoff = 0.05,
  get_position = NULL,
  get_icon = NULL,
  get_color = NULL,
  get_size = NULL,
  get_angle = NULL,
  ...
)
}
\arguments{
\item{rdeck}{\\verb{\{rdeck\}\\} an rdeck widget instance}

\item{data}{\code{{data.frame | sf}}}

\item{visible}{\code{{logical}}}

\item{pickable}{\code{{logical}}}

\item{opacity}{\code{{numeric}}}

\item{position_format}{\code{{"XY" | "XYZ"}}}

\item{color_format}{\code{{"RGB" | "RGBA"}}}

\item{auto_highlight}{\code{{logical}}}

\item{highlight_color}{\code{{integer}}}

\item{icon_atlas}{\code{{list}}}

\item{icon_mapping}{\code{{list}}}

\item{size_scale}{\code{{numeric}}}

\item{billboard}{\code{{logical}}}

\item{size_units}{\code{{"pixels" | "meters"}}}

\item{size_min_pixels}{\code{{numeric}}}

\item{size_max_pixels}{\code{{numeric}}}

\item{alpha_cutoff}{\code{{numeric}}}

\item{get_position}{\code{{accessor | JS}}}

\item{get_icon}{\code{{accessor | JS}}}

\item{get_color}{\code{{accessor}}}

\item{get_size}{\code{{accessor | numeric}}}

\item{get_angle}{\code{{accessor | numeric}}}

\item{...}{additional layer parameters to pass to deck.gl}
}
\value{
\\verb{\{rdeck\}\\}
}
\description{
Add IconLayer to an rdeck map.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/icon-layer.md}
}
