% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_layer.R
\name{add_layer}
\alias{add_layer}
\title{add_layer}
\usage{
add_layer(rdeck, layer)
}
\arguments{
\item{rdeck}{\code{{rdeck}}}

\item{layer}{\code{{layer}}}
}
\description{
Add layer to deck
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/master/docs/api-reference/layer.md}
}
