% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_line_layer.R
\name{add_line_layer}
\alias{add_line_layer}
\title{Add LineLayer to an rdeck map.}
\usage{
add_line_layer(
  rdeck,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  get_source_position = NULL,
  get_target_position = NULL,
  get_color = NULL,
  get_width = NULL,
  width_units = "pixels",
  width_scale = 1,
  width_min_pixels = 0,
  width_max_pixels = 9007199254740991,
  ...
)
}
\arguments{
\item{rdeck}{\\verb{\{rdeck\}\\} an rdeck widget instance}

\item{data}{\code{{data.frame | sf}}}

\item{visible}{\code{{logical}}}

\item{pickable}{\code{{logical}}}

\item{opacity}{\code{{numeric}}}

\item{position_format}{\code{{"XY" | "XYZ"}}}

\item{color_format}{\code{{"RGB" | "RGBA"}}}

\item{auto_highlight}{\code{{logical}}}

\item{highlight_color}{\code{{integer}}}

\item{get_source_position}{\code{{accessor | JS}}}

\item{get_target_position}{\code{{accessor | JS}}}

\item{get_color}{\code{{accessor}}}

\item{get_width}{\code{{accessor | numeric}}}

\item{width_units}{\code{{"pixels" | "meters"}}}

\item{width_scale}{\code{{numeric}}}

\item{width_min_pixels}{\code{{numeric}}}

\item{width_max_pixels}{\code{{numeric}}}

\item{...}{additional layer parameters to pass to deck.gl}
}
\value{
\\verb{\{rdeck\}\\}
}
\description{
Add LineLayer to an rdeck map.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/line-layer.md}
}
