% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_polygon_layer.R
\name{add_polygon_layer}
\alias{add_polygon_layer}
\title{Add PolygonLayer to an rdeck map.}
\usage{
add_polygon_layer(
  rdeck,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  stroked = TRUE,
  filled = TRUE,
  extruded = FALSE,
  elevation_scale = 1,
  wireframe = FALSE,
  line_width_units = "meters",
  line_width_scale = 1,
  line_width_min_pixels = 0,
  line_width_max_pixels = 9007199254740991,
  line_joint_rounded = FALSE,
  line_miter_limit = 4,
  get_polygon = NULL,
  get_fill_color = NULL,
  get_line_color = NULL,
  get_line_width = NULL,
  get_elevation = NULL,
  material = TRUE,
  ...
)
}
\arguments{
\item{rdeck}{\\verb{\{rdeck\}\\} an rdeck widget instance}

\item{data}{\code{{data.frame | sf}}}

\item{visible}{\code{{logical}}}

\item{pickable}{\code{{logical}}}

\item{opacity}{\code{{numeric}}}

\item{position_format}{\code{{"XY" | "XYZ"}}}

\item{color_format}{\code{{"RGB" | "RGBA"}}}

\item{auto_highlight}{\code{{logical}}}

\item{highlight_color}{\code{{integer}}}

\item{stroked}{\code{{logical}}}

\item{filled}{\code{{logical}}}

\item{extruded}{\code{{logical}}}

\item{elevation_scale}{\code{{numeric}}}

\item{wireframe}{\code{{logical}}}

\item{line_width_units}{\code{{"pixels" | "meters"}}}

\item{line_width_scale}{\code{{numeric}}}

\item{line_width_min_pixels}{\code{{numeric}}}

\item{line_width_max_pixels}{\code{{numeric}}}

\item{line_joint_rounded}{\code{{logical}}}

\item{line_miter_limit}{\code{{numeric}}}

\item{get_polygon}{\code{{accessor | JS}}}

\item{get_fill_color}{\code{{accessor}}}

\item{get_line_color}{\code{{accessor}}}

\item{get_line_width}{\code{{accessor | numeric}}}

\item{get_elevation}{\code{{accessor | numeric}}}

\item{material}{\code{{logical}}}

\item{...}{additional layer parameters to pass to deck.gl}
}
\value{
\\verb{\{rdeck\}\\}
}
\description{
Add PolygonLayer to an rdeck map.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/polygon-layer.md}
}
