% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_screen_grid_layer.R
\name{add_screen_grid_layer}
\alias{add_screen_grid_layer}
\title{Add ScreenGridLayer to an rdeck map.}
\usage{
add_screen_grid_layer(
  rdeck,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  cell_size_pixels = 100,
  cell_margin_pixels = 2,
  color_domain = NULL,
  color_range = list(c(255, 255, 178), c(254, 217, 118), c(254, 178, 76), c(253, 141,
    60), c(240, 59, 32), c(189, 0, 38)),
  get_position = NULL,
  get_weight = NULL,
  gpu_aggregation = TRUE,
  aggregation = "SUM",
  ...
)
}
\arguments{
\item{rdeck}{\\verb{\{rdeck\}\\} an rdeck widget instance}

\item{data}{\code{{data.frame | sf}}}

\item{visible}{\code{{logical}}}

\item{pickable}{\code{{logical}}}

\item{opacity}{\code{{numeric}}}

\item{position_format}{\code{{"XY" | "XYZ"}}}

\item{color_format}{\code{{"RGB" | "RGBA"}}}

\item{auto_highlight}{\code{{logical}}}

\item{highlight_color}{\code{{integer}}}

\item{cell_size_pixels}{\code{{numeric}}}

\item{cell_margin_pixels}{\code{{numeric}}}

\item{color_domain}{\code{{numeric}}}

\item{color_range}{\code{{list}}}

\item{get_position}{\code{{accessor | JS}}}

\item{get_weight}{\code{{accessor | JS}}}

\item{gpu_aggregation}{\code{{logical}}}

\item{aggregation}{\code{{"SUM" | "MEAN" | "MIN" | "MAX"}}}

\item{...}{additional layer parameters to pass to deck.gl}
}
\value{
\\verb{\{rdeck\}\\}
}
\description{
Add ScreenGridLayer to an rdeck map.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/screen-grid-layer.md}
}
