% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_simple_mesh_layer.R
\name{add_simple_mesh_layer}
\alias{add_simple_mesh_layer}
\title{Add SimpleMeshLayer to an rdeck map.}
\usage{
add_simple_mesh_layer(
  rdeck,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  mesh = NULL,
  texture = NULL,
  size_scale = 1,
  wireframe = FALSE,
  material = TRUE,
  get_position = NULL,
  get_color = NULL,
  get_orientation = NULL,
  get_scale = NULL,
  get_translation = NULL,
  get_transform_matrix = NULL,
  ...
)
}
\arguments{
\item{rdeck}{\\verb{\{rdeck\}\\} an rdeck widget instance}

\item{data}{\code{{data.frame | sf}}}

\item{visible}{\code{{logical}}}

\item{pickable}{\code{{logical}}}

\item{opacity}{\code{{numeric}}}

\item{position_format}{\code{{"XY" | "XYZ"}}}

\item{color_format}{\code{{"RGB" | "RGBA"}}}

\item{auto_highlight}{\code{{logical}}}

\item{highlight_color}{\code{{integer}}}

\item{mesh}{\code{{list}}}

\item{texture}{\code{{list}}}

\item{size_scale}{\code{{numeric}}}

\item{wireframe}{\code{{logical}}}

\item{material}{\code{{logical}}}

\item{get_position}{\code{{accessor | JS}}}

\item{get_color}{\code{{accessor}}}

\item{get_orientation}{\code{{accessor}}}

\item{get_scale}{\code{{accessor}}}

\item{get_translation}{\code{{accessor}}}

\item{get_transform_matrix}{\code{{accessor}}}

\item{...}{additional layer parameters to pass to deck.gl}
}
\value{
\\verb{\{rdeck\}\\}
}
\description{
Add SimpleMeshLayer to an rdeck map.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/simple-mesh-layer.md}
}
