% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_text_layer.R
\name{add_text_layer}
\alias{add_text_layer}
\title{Add TextLayer to an rdeck map.}
\usage{
add_text_layer(
  rdeck,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  billboard = TRUE,
  size_scale = 1,
  size_units = "pixels",
  size_min_pixels = 0,
  size_max_pixels = 9007199254740991,
  background_color = NULL,
  character_set = NULL,
  font_family = "Monaco, monospace",
  font_weight = "normal",
  line_height = 1,
  font_settings = NULL,
  word_break = "word-break",
  max_width = -1,
  get_text = NULL,
  get_position = NULL,
  get_color = NULL,
  get_size = NULL,
  get_angle = NULL,
  get_text_anchor = NULL,
  get_alignment_baseline = NULL,
  get_pixel_offset = NULL,
  ...
)
}
\arguments{
\item{rdeck}{\\verb{\{rdeck\}\\} an rdeck widget instance}

\item{data}{\code{{data.frame | sf}}}

\item{visible}{\code{{logical}}}

\item{pickable}{\code{{logical}}}

\item{opacity}{\code{{numeric}}}

\item{position_format}{\code{{"XY" | "XYZ"}}}

\item{color_format}{\code{{"RGB" | "RGBA"}}}

\item{auto_highlight}{\code{{logical}}}

\item{highlight_color}{\code{{integer}}}

\item{billboard}{\code{{logical}}}

\item{size_scale}{\code{{numeric}}}

\item{size_units}{\code{{"pixels" | "meters"}}}

\item{size_min_pixels}{\code{{numeric}}}

\item{size_max_pixels}{\code{{numeric}}}

\item{background_color}{\code{{integer}}}

\item{character_set}{\code{{character}}}

\item{font_family}{\code{{character}}}

\item{font_weight}{\code{{character}}}

\item{line_height}{\code{{numeric}}}

\item{font_settings}{\code{{list}}}

\item{word_break}{\code{{"break-all" | "break-word"}}}

\item{max_width}{\code{{numeric}}}

\item{get_text}{\code{{accessor | JS}}}

\item{get_position}{\code{{accessor | JS}}}

\item{get_color}{\code{{accessor}}}

\item{get_size}{\code{{accessor | numeric}}}

\item{get_angle}{\code{{accessor | numeric}}}

\item{get_text_anchor}{\code{{accessor | character}}}

\item{get_alignment_baseline}{\code{{accessor | character}}}

\item{get_pixel_offset}{\code{{accessor}}}

\item{...}{additional layer parameters to pass to deck.gl}
}
\value{
\\verb{\{rdeck\}\\}
}
\description{
Add TextLayer to an rdeck map.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/text-layer.md}
}
