% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{bounds}
\alias{bounds}
\title{bounds}
\usage{
bounds(obj)
}
\arguments{
\item{obj}{\code{{sf | sfc | sfg}}
A simple features object, column or geometry to compute bounds from}
}
\value{
\code{{bounds}}
}
\description{
Create a bounds instance from a simple features object, column or geometry
}
