% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_breaks.R
\name{breaks_log}
\alias{breaks_log}
\alias{log_breaks}
\title{Breaks log}
\usage{
breaks_log(n = 10, base = exp(1))
}
\arguments{
\item{n}{<\code{int}> the size of the output vector. Ouput size will be at least length-2
for finite input.}

\item{base}{<\code{number}> The log base}
}
\description{
Generate a breaks vector of size \code{n} with log \link{log_trans} spaced breaks.
}
\details{
Input range must be finite and either be strictly negative or strictly positive;
it must not cross, nor include 0.

If input range is non-finite, an empty vector is returned.
}
\note{
This function has a different goal to \link[scales:breaks_log]{scales::breaks_log}: it produces evenly
spaced log-breaks for use with d3-scale, it doesn't produce breaks with pretty
labels.
}
\examples{
breaks_log(5)(-10:-1)
breaks_log(5)(1:10)

}
\seealso{
Other breaks: 
\code{\link{breaks_linear}()},
\code{\link{breaks_manual}()},
\code{\link{breaks_power}()},
\code{\link{breaks_symlog}()},
\code{\link{breaks_trans}()}
}
\concept{breaks}
