% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_breaks.R
\name{breaks_manual}
\alias{breaks_manual}
\alias{manual_breaks}
\title{Breaks manual}
\usage{
breaks_manual(thresholds = 0.5)
}
\arguments{
\item{thresholds}{<\code{numeric}> a numeric vector of \emph{ordered}, finite thresholds.}
}
\description{
Generate a breaks vector with pre-determined threshold values.
}
\details{
Wraps \code{thresholds} in the range of input data (which may be null or non-finite).
Thresholds that fall outside the range of input are omitted from the output.
}
\examples{
breaks_manual(0)(-10:10)
breaks_manual(c(-10, 0, 10))(-10:10)
breaks_manual(-1:1 * 1e6)(NULL)

}
\seealso{
Other breaks: 
\code{\link{breaks_linear}()},
\code{\link{breaks_log}()},
\code{\link{breaks_power}()},
\code{\link{breaks_symlog}()},
\code{\link{breaks_trans}()}
}
\concept{breaks}
