% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_breaks.R
\name{breaks_power}
\alias{breaks_power}
\alias{power_breaks}
\title{Breaks power}
\usage{
breaks_power(n = 10, exponent = 0.5)
}
\arguments{
\item{n}{<\code{int}> the size of the output vector. Ouput size will be at least length-2
for finite input.}

\item{exponent}{<\code{number}> The power exponent}
}
\description{
Generate a breaks vector of size \code{n} with exponentially \link{power_trans} spaced breaks.
}
\details{
If input range is non-finite, an empty vector is returned.
}
\examples{
breaks_power(5)(-10:10)
breaks_power(5, exponent = 1 / 3)(-1:1)

}
\seealso{
Other breaks: 
\code{\link{breaks_linear}()},
\code{\link{breaks_log}()},
\code{\link{breaks_manual}()},
\code{\link{breaks_symlog}()},
\code{\link{breaks_trans}()}
}
\concept{breaks}
