% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_breaks.R
\name{breaks_symlog}
\alias{breaks_symlog}
\alias{symlog_breaks}
\title{Breaks symlog}
\usage{
breaks_symlog(n = 10)
}
\arguments{
\item{n}{<\code{int}> the size of the output vector. Ouput size will be at least length-2
for finite input.}
}
\description{
Generate a breaks vector of size \code{n} with log1p \link{symlog_trans} spaced breaks.
}
\details{
If input range is non-finite, an empty vector is returned.
}
\examples{
breaks_symlog(5)(-10:10)
breaks_symlog(5)(0:10)
}
\seealso{
Other breaks: 
\code{\link{breaks_linear}()},
\code{\link{breaks_log}()},
\code{\link{breaks_manual}()},
\code{\link{breaks_power}()},
\code{\link{breaks_trans}()}
}
\concept{breaks}
