% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_breaks.R
\name{breaks_trans}
\alias{breaks_trans}
\alias{trans_breaks}
\title{Breaks trans}
\usage{
breaks_trans(n = 10, trans)
}
\arguments{
\item{n}{<\code{int}> the size of the output vector. Ouput size will be at least length-2
for finite input.}

\item{trans}{<\code{scales::trans}> an \emph{invertible} transformer.}
}
\description{
Generate a breaks vector of size \code{n} with evenly spaced breaks in \code{trans} domain.
}
\details{
Breaks are generated by transforming the input range to \code{trans} domain, generating
a regular sequence of size \code{n} for the transformed range, then inverting the transform.

The input range must be finite for both the input domain \emph{and} the \code{trans} domain. If
input is not finite, an empty vector is returned.
}
\examples{
breaks_trans(trans = scales::identity_trans())(-10:10)
breaks_trans(trans = symlog_trans())(-10:10)
breaks_trans(trans = power_trans())(-10:10)
breaks_manual(c(-10, 0, 10))(-10:10)
breaks_manual(-1:1 * 1e6)(NULL)

}
\seealso{
Other breaks: 
\code{\link{breaks_linear}()},
\code{\link{breaks_log}()},
\code{\link{breaks_manual}()},
\code{\link{breaks_power}()},
\code{\link{breaks_symlog}()}
}
\concept{breaks}
