% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-proptypes.R
\name{color}
\alias{color}
\title{Prop Type: Color}
\description{
Colours are represented by \code{RGB} or \code{RGBA} hex strings.
Example: \verb{#663399ff} \if{html}{\out{<img width="20" src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVQYV2NIM575HwAEZwIytoDTwwAAAABJRU5ErkJggg==" />}}
}
\keyword{internal}
