% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cur_value}
\alias{cur_value}
\title{Cur value}
\usage{
cur_value()
}
\description{
A sentinel object that represents the current value in the browser. Use this value
to indicate that a map / layer property should remain unchanged.

Intended for use in shiny applications.
}
\examples{
rdeck_proxy("map") \%>\%
  set_layer_visibility("layer_id", visible = cur_value(), visibility_toggle = TRUE)

}
\keyword{internal}
