% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{editor_options}
\alias{editor_options}
\title{Polygon editor options}
\usage{
editor_options(mode = cur_value(), features = cur_value())
}
\arguments{
\item{mode}{<\code{"view"} | \code{"modify"} | \code{"polygon"} | \code{"lasso"}> The polygon editor mode.
\itemize{
\item \code{view}: editor is in readonly mode
\item \code{modify}: add/move/delete vertices, or move the entire polygon
\item \code{polygon}: draw a polygon by clicking each vertex
\item \code{lasso}: freehand polygon draw by click-dragging
}}

\item{features}{<\code{sfc}> Features with which to initialise the editor}
}
\description{
Options for the polygon editor
}
