% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{format_number}
\alias{format_number}
\title{Format number}
\usage{
format_number(tick, digits = 2)
}
\arguments{
\item{tick}{a vector of tick values}

\item{digits}{the number of digits to output}
}
\description{
Format numeric and integer values.
}
