% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{get_bounds}
\alias{get_bounds}
\title{get_bounds}
\usage{
get_bounds(obj)
}
\arguments{
\item{obj}{\code{{sf | sfc | sfg}}}
}
\value{
\code{{MULTIPOINT}}
}
\description{
Convert a bbox instance into a multipoint geometry
}
