% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{get_clicked_layer}
\alias{get_clicked_layer}
\title{Get clicked layer}
\usage{
get_clicked_layer(rdeck, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{rdeck}{<\code{rdeck_proxy} | \code{string}> the map, or map id}

\item{session}{<\code{ShinySession}> the shiny session}
}
\description{
Get the clicked layer
}
