% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{get_event_data}
\alias{get_event_data}
\title{Get event data}
\usage{
get_event_data(rdeck, event_name, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{rdeck}{<\code{rdeck_proxy} | \code{string}> the map, or map id}

\item{event_name}{<\code{string}> the event name}

\item{session}{<\code{ShinySession}> the shiny session}
}
\description{
Get event data
}
\keyword{internal}
