% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{get_position_format}
\alias{get_position_format}
\title{get_position_format}
\usage{
get_position_format(obj)
}
\arguments{
\item{obj}{\code{{sf | sfc | sfg | data.frame}}}
}
\value{
\code{{XY|XYZ}}
}
\description{
Get position format of geometry
}
