% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{get_view_bounds}
\alias{get_view_bounds}
\title{Get view bounds}
\usage{
get_view_bounds(rdeck, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{rdeck}{<\code{rdeck_proxy} | \code{string}> the map, or map id}

\item{session}{<\code{ShinySession}> the shiny session}
}
\description{
Get the view bounds
}
