% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{get_view_state}
\alias{get_view_state}
\title{Get view state}
\usage{
get_view_state(rdeck, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{rdeck}{<\code{rdeck_proxy} | \code{string}> the map, or map id}

\item{session}{<\code{ShinySession}> the shiny session}
}
\description{
Get the view state
}
