% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/great_circle_layer.R
\name{great_circle_layer}
\alias{great_circle_layer}
\title{\href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/great-circle-layer.md}{GreatCircleLayer} deck.gl layer.}
\usage{
great_circle_layer(
  id = NULL,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  get_source_position = NULL,
  get_target_position = NULL,
  get_source_color = NULL,
  get_target_color = NULL,
  get_width = NULL,
  get_height = NULL,
  get_tilt = NULL,
  width_units = "pixels",
  width_scale = 1,
  width_min_pixels = 0,
  width_max_pixels = 9007199254740991,
  ...
)
}
\arguments{
\item{id}{\code{\link{character}}
The id of the layer. Layer ids must be unique per layer \code{type} for deck.gl
to properly distinguish between them.}

\item{data}{\code{\link{data.frame}} | \code{\link[sf:sf]{sf::sf}}}

\item{visible}{\code{\link{logical}}}

\item{pickable}{\code{\link{logical}}}

\item{opacity}{\code{\link{numeric}}}

\item{position_format}{\code{XY} | \code{XYZ}}

\item{color_format}{\code{RGB} | \code{RGBA}}

\item{auto_highlight}{\code{\link{logical}}}

\item{highlight_color}{\code{\link{integer}}}

\item{get_source_position}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{get_target_position}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{get_source_color}{accessor}

\item{get_target_color}{accessor}

\item{get_width}{accessor | \code{\link{numeric}}}

\item{get_height}{accessor | \code{\link{numeric}}}

\item{get_tilt}{accessor | \code{\link{numeric}}}

\item{width_units}{\code{pixels} | \code{meters}}

\item{width_scale}{\code{\link{numeric}}}

\item{width_min_pixels}{\code{\link{numeric}}}

\item{width_max_pixels}{\code{\link{numeric}}}

\item{...}{additional layer parameters to pass to deck.gl.
\code{snake_case} parameters will be converted to \code{snakeCase}.}
}
\value{
\code{GreatCircleLayer} & \code{\link{layer}}
A \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/great-circle-layer.md}{GreatCircleLayer} layer.
Add to an \link{rdeck} map via \code{\link{add_layer}} or \code{\link{rdeck}}.
}
\description{
\href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/great-circle-layer.md}{GreatCircleLayer} deck.gl layer.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/great-circle-layer.md}
}
