% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deckgl-layers-doc.R, R/deckgl-layers.R
\name{grid_layer}
\alias{grid_layer}
\alias{add_grid_layer}
\title{Grid Layer}
\usage{
add_grid_layer(
  rdeck,
  ...,
  id = NULL,
  name = "GridLayer",
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  color_domain = NULL,
  color_range = c("#ffffb2", "#fed976", "#feb24c", "#fd8d3c", "#f03b20", "#bd0026"),
  get_color_weight = 1,
  color_aggregation = "SUM",
  elevation_domain = NULL,
  elevation_range = c(0, 1000),
  get_elevation_weight = 1,
  elevation_aggregation = "SUM",
  elevation_scale = 1,
  cell_size = 1000,
  coverage = 1,
  get_position = position,
  extruded = FALSE,
  material = TRUE,
  get_color_value = NULL,
  lower_percentile = 0,
  upper_percentile = 100,
  color_scale_type = "quantize",
  get_elevation_value = NULL,
  elevation_lower_percentile = 0,
  elevation_upper_percentile = 100,
  elevation_scale_type = "linear",
  grid_aggregator = NULL,
  gpu_aggregation = FALSE,
  tooltip = FALSE
)
}
\arguments{
\item{id}{string}

\item{name}{string}

\item{data}{data}

\item{visible}{boolean}

\item{pickable}{boolean}

\item{opacity}{number}

\item{position_format}{unknown}

\item{color_format}{unknown}

\item{auto_highlight}{boolean}

\item{highlight_color}{accessor}

\item{color_domain}{unknown}

\item{color_range}{array}

\item{get_color_weight}{accessor}

\item{color_aggregation}{unknown}

\item{elevation_domain}{unknown}

\item{elevation_range}{array}

\item{get_elevation_weight}{accessor}

\item{elevation_aggregation}{unknown}

\item{elevation_scale}{number}

\item{cell_size}{number}

\item{coverage}{number}

\item{get_position}{accessor}

\item{extruded}{boolean}

\item{material}{boolean}

\item{get_color_value}{unknown}

\item{lower_percentile}{number}

\item{upper_percentile}{number}

\item{color_scale_type}{unknown}

\item{get_elevation_value}{unknown}

\item{elevation_lower_percentile}{number}

\item{elevation_upper_percentile}{number}

\item{elevation_scale_type}{unknown}

\item{grid_aggregator}{function}

\item{gpu_aggregation}{boolean}

\item{tooltip}{accessor}
}
\description{
Grid Layer
}
\seealso{
\url{https://github.com/visgl/deck.gl/blob/v8.3.11/docs/api-reference/aggregation-layers/grid-layer.md}

Other layers: 
\code{\link{arc_layer}},
\code{\link{bitmap_layer}},
\code{\link{column_layer}},
\code{\link{contour_layer}},
\code{\link{cpu_grid_layer}},
\code{\link{geojson_layer}},
\code{\link{gpu_grid_layer}},
\code{\link{great_circle_layer}},
\code{\link{grid_cell_layer}},
\code{\link{h3_cluster_layer}},
\code{\link{h3_hexagon_layer}},
\code{\link{heatmap_layer}},
\code{\link{hexagon_layer}},
\code{\link{icon_layer}},
\code{\link{line_layer}},
\code{\link{mvt_layer}},
\code{\link{path_layer}},
\code{\link{point_cloud_layer}},
\code{\link{polygon_layer}},
\code{\link{s2_layer}},
\code{\link{scatterplot_layer}},
\code{\link{screen_grid_layer}},
\code{\link{simple_mesh_layer}},
\code{\link{solid_polygon_layer}},
\code{\link{terrain_layer}},
\code{\link{text_layer}},
\code{\link{tile_3d_layer}},
\code{\link{tile_layer}},
\code{\link{trips_layer}}

Other aggregation-layers: 
\code{\link{contour_layer}},
\code{\link{cpu_grid_layer}},
\code{\link{gpu_grid_layer}},
\code{\link{heatmap_layer}},
\code{\link{hexagon_layer}},
\code{\link{screen_grid_layer}}
}
\concept{aggregation-layers}
\concept{layers}
