% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3_cluster_layer.R
\name{h3_cluster_layer}
\alias{h3_cluster_layer}
\title{\href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/h3-cluster-layer.md}{H3ClusterLayer} deck.gl layer.}
\usage{
h3_cluster_layer(
  id = NULL,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  get_hexagons = NULL,
  stroked = TRUE,
  filled = TRUE,
  extruded = FALSE,
  elevation_scale = 1,
  wireframe = FALSE,
  line_width_units = "meters",
  line_width_scale = 1,
  line_width_min_pixels = 0,
  line_width_max_pixels = 9007199254740991,
  line_joint_rounded = FALSE,
  line_miter_limit = 4,
  get_polygon = NULL,
  get_fill_color = NULL,
  get_line_color = NULL,
  get_line_width = NULL,
  get_elevation = NULL,
  material = TRUE,
  ...
)
}
\arguments{
\item{id}{\code{\link{character}}
The id of the layer. Layer ids must be unique per layer \code{type} for deck.gl
to properly distinguish between them.}

\item{data}{\code{\link{data.frame}} | \code{\link[sf:sf]{sf::sf}}}

\item{visible}{\code{\link{logical}}}

\item{pickable}{\code{\link{logical}}}

\item{opacity}{\code{\link{numeric}}}

\item{position_format}{\code{XY} | \code{XYZ}}

\item{color_format}{\code{RGB} | \code{RGBA}}

\item{auto_highlight}{\code{\link{logical}}}

\item{highlight_color}{\code{\link{integer}}}

\item{get_hexagons}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{stroked}{\code{\link{logical}}}

\item{filled}{\code{\link{logical}}}

\item{extruded}{\code{\link{logical}}}

\item{elevation_scale}{\code{\link{numeric}}}

\item{wireframe}{\code{\link{logical}}}

\item{line_width_units}{\code{pixels} | \code{meters}}

\item{line_width_scale}{\code{\link{numeric}}}

\item{line_width_min_pixels}{\code{\link{numeric}}}

\item{line_width_max_pixels}{\code{\link{numeric}}}

\item{line_joint_rounded}{\code{\link{logical}}}

\item{line_miter_limit}{\code{\link{numeric}}}

\item{get_polygon}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{get_fill_color}{accessor}

\item{get_line_color}{accessor}

\item{get_line_width}{accessor | \code{\link{numeric}}}

\item{get_elevation}{accessor | \code{\link{numeric}}}

\item{material}{\code{\link{logical}}}

\item{...}{additional layer parameters to pass to deck.gl.
\code{snake_case} parameters will be converted to \code{snakeCase}.}
}
\value{
\code{H3ClusterLayer} & \code{\link{layer}}
A \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/h3-cluster-layer.md}{H3ClusterLayer} layer.
Add to an \link{rdeck} map via \code{\link{add_layer}} or \code{\link{rdeck}}.
}
\description{
\href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/h3-cluster-layer.md}{H3ClusterLayer} deck.gl layer.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/h3-cluster-layer.md}
}
