% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon_layer.R
\name{icon_layer}
\alias{icon_layer}
\title{\href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/icon-layer.md}{IconLayer} deck.gl layer.}
\usage{
icon_layer(
  id = NULL,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  icon_atlas = NULL,
  icon_mapping = NULL,
  size_scale = 1,
  billboard = TRUE,
  size_units = "pixels",
  size_min_pixels = 0,
  size_max_pixels = 9007199254740991,
  alpha_cutoff = 0.05,
  get_position = NULL,
  get_icon = NULL,
  get_color = NULL,
  get_size = NULL,
  get_angle = NULL,
  ...
)
}
\arguments{
\item{id}{\code{\link{character}}
The id of the layer. Layer ids must be unique per layer \code{type} for deck.gl
to properly distinguish between them.}

\item{data}{\code{\link{data.frame}} | \code{\link[sf:sf]{sf::sf}}}

\item{visible}{\code{\link{logical}}}

\item{pickable}{\code{\link{logical}}}

\item{opacity}{\code{\link{numeric}}}

\item{position_format}{\code{XY} | \code{XYZ}}

\item{color_format}{\code{RGB} | \code{RGBA}}

\item{auto_highlight}{\code{\link{logical}}}

\item{highlight_color}{\code{\link{integer}}}

\item{icon_atlas}{\code{\link{list}}}

\item{icon_mapping}{\code{\link{list}}}

\item{size_scale}{\code{\link{numeric}}}

\item{billboard}{\code{\link{logical}}}

\item{size_units}{\code{pixels} | \code{meters}}

\item{size_min_pixels}{\code{\link{numeric}}}

\item{size_max_pixels}{\code{\link{numeric}}}

\item{alpha_cutoff}{\code{\link{numeric}}}

\item{get_position}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{get_icon}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{get_color}{accessor}

\item{get_size}{accessor | \code{\link{numeric}}}

\item{get_angle}{accessor | \code{\link{numeric}}}

\item{...}{additional layer parameters to pass to deck.gl.
\code{snake_case} parameters will be converted to \code{camelCase}.}
}
\value{
\code{IconLayer} & \code{\link{layer}}
A \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/icon-layer.md}{IconLayer} layer.
Add to an \link{rdeck} map via \code{\link{add_layer}} or \code{\link{rdeck}}.
}
\description{
\href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/icon-layer.md}{IconLayer} deck.gl layer.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/icon-layer.md}
}
