% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{layer_data}
\alias{layer_data}
\title{Get layer data}
\usage{
layer_data(data, layer_type)
}
\arguments{
\item{data}{\code{\link{data.frame}} | \code{\link[sf:sf]{sf::sf}}}
}
\value{
\code{\link{list}} | \code{json}
}
\description{
Get layer data
}
\keyword{internal}
