% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_types.R
\docType{data}
\name{layer_types}
\alias{layer_types}
\title{Supported deck.gl layer types}
\format{An object of class \code{character} of length 29.}
\usage{
layer_types
}
\description{
Supported deck.gl layer types
}
\keyword{datasets}
