% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{mapbox_access_token}
\alias{mapbox_access_token}
\title{Mapbox access token}
\usage{
mapbox_access_token()
}
\description{
A mapbox access token is required for rendering the mapbox basemap (regardless of tiles used)
and mapbox services (tiles). To use a basemap, you need to register for a
\href{https://account.mapbox.com/auth/signup}{mapbox account}. Mapbox has a generous free tier.

Each rdeck map \emph{rendered} equates to a
\href{https://www.mapbox.com/pricing/#maploads}{map load for web}.
}
\details{
The mapbox token is read from the following locations (in order):
\itemize{
\item \code{getOption("rdeck.mapbox_access_token")}
\item \code{Sys.getenv("MAPBOX_ACCESS_TOKEN")}
\item `Sys.getenv("MAPBOX_TOKEN")
}
}
\seealso{
\url{https://docs.mapbox.com/help/glossary/access-token}
}
