% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_json.R
\name{mvt_url}
\alias{mvt_url}
\title{MVT URL}
\usage{
mvt_url(tileset_id)
}
\arguments{
\item{tileset_id}{A mapbox tileset identifier of the form:
\itemize{
\item \code{mapbox.mapbox-streets-v8}, or
\item \verb{mapbox://mapbox.mapbox-streets-v8}
}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Please url \code{\link[=tile_json]{tile_json()}} instead
}
\keyword{internal}
