% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_cloud_layer.R
\name{point_cloud_layer}
\alias{point_cloud_layer}
\title{\href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/point-cloud-layer.md}{PointCloudLayer} deck.gl layer.}
\usage{
point_cloud_layer(
  id = NULL,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  size_units = "pixels",
  point_size = 10,
  get_position = NULL,
  get_normal = NULL,
  get_color = NULL,
  material = TRUE,
  ...
)
}
\arguments{
\item{id}{\code{\link{character}}
The id of the layer. Layer ids must be unique per layer \code{type} for deck.gl
to properly distinguish between them.}

\item{data}{\code{\link{data.frame}} | \code{\link[sf:sf]{sf::sf}}}

\item{visible}{\code{\link{logical}}}

\item{pickable}{\code{\link{logical}}}

\item{opacity}{\code{\link{numeric}}}

\item{position_format}{\code{XY} | \code{XYZ}}

\item{color_format}{\code{RGB} | \code{RGBA}}

\item{auto_highlight}{\code{\link{logical}}}

\item{highlight_color}{\code{\link{integer}}}

\item{size_units}{\code{pixels} | \code{meters}}

\item{point_size}{\code{\link{numeric}}}

\item{get_position}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{get_normal}{accessor}

\item{get_color}{accessor}

\item{material}{\code{\link{logical}}}

\item{...}{additional layer parameters to pass to deck.gl.
\code{snake_case} parameters will be converted to \code{snakeCase}.}
}
\value{
\code{PointCloudLayer} & \code{\link{layer}}
A \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/point-cloud-layer.md}{PointCloudLayer} layer.
Add to an \link{rdeck} map via \code{\link{add_layer}} or \code{\link{rdeck}}.
}
\description{
\href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/point-cloud-layer.md}{PointCloudLayer} deck.gl layer.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/point-cloud-layer.md}
}
