% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deckgl-layers-doc.R, R/deckgl-layers.R
\name{polygon_layer}
\alias{polygon_layer}
\alias{add_polygon_layer}
\title{Polygon Layer}
\usage{
add_polygon_layer(
  rdeck,
  ...,
  id = NULL,
  name = "PolygonLayer",
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  stroked = TRUE,
  filled = TRUE,
  extruded = FALSE,
  elevation_scale = 1,
  wireframe = FALSE,
  line_width_units = "meters",
  line_width_scale = 1,
  line_width_min_pixels = 0,
  line_width_max_pixels = 9007199254740991,
  line_joint_rounded = FALSE,
  line_miter_limit = 4,
  get_polygon = polygon,
  get_fill_color = "#000000ff",
  get_line_color = "#000000ff",
  get_line_width = 1,
  get_elevation = 1000,
  material = TRUE,
  tooltip = FALSE
)
}
\arguments{
\item{id}{string}

\item{name}{string}

\item{data}{data}

\item{visible}{boolean}

\item{pickable}{boolean}

\item{opacity}{number}

\item{position_format}{unknown}

\item{color_format}{unknown}

\item{auto_highlight}{boolean}

\item{highlight_color}{accessor}

\item{stroked}{boolean}

\item{filled}{boolean}

\item{extruded}{boolean}

\item{elevation_scale}{number}

\item{wireframe}{boolean}

\item{line_width_units}{unknown}

\item{line_width_scale}{number}

\item{line_width_min_pixels}{number}

\item{line_width_max_pixels}{number}

\item{line_joint_rounded}{boolean}

\item{line_miter_limit}{number}

\item{get_polygon}{accessor}

\item{get_fill_color}{accessor}

\item{get_line_color}{accessor}

\item{get_line_width}{accessor}

\item{get_elevation}{accessor}

\item{material}{boolean}

\item{tooltip}{accessor}
}
\description{
Polygon Layer
}
\seealso{
\url{https://github.com/visgl/deck.gl/blob/v8.3.11/docs/api-reference/layers/polygon-layer.md}

Other layers: 
\code{\link{arc_layer}},
\code{\link{bitmap_layer}},
\code{\link{column_layer}},
\code{\link{contour_layer}},
\code{\link{cpu_grid_layer}},
\code{\link{geojson_layer}},
\code{\link{gpu_grid_layer}},
\code{\link{great_circle_layer}},
\code{\link{grid_cell_layer}},
\code{\link{grid_layer}},
\code{\link{h3_cluster_layer}},
\code{\link{h3_hexagon_layer}},
\code{\link{heatmap_layer}},
\code{\link{hexagon_layer}},
\code{\link{icon_layer}},
\code{\link{line_layer}},
\code{\link{mvt_layer}},
\code{\link{path_layer}},
\code{\link{point_cloud_layer}},
\code{\link{s2_layer}},
\code{\link{scatterplot_layer}},
\code{\link{screen_grid_layer}},
\code{\link{simple_mesh_layer}},
\code{\link{solid_polygon_layer}},
\code{\link{terrain_layer}},
\code{\link{text_layer}},
\code{\link{tile_3d_layer}},
\code{\link{tile_layer}},
\code{\link{trips_layer}}
}
\concept{layers}
