% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_trans.R
\name{power_trans}
\alias{power_trans}
\title{Power transformation}
\usage{
power_trans(exponent = 0.5)
}
\arguments{
\item{exponent}{<\code{number}> The power exponent}
}
\description{
Applies an exponential transform of the input: \code{y = abs(x)^k}, where:
\itemize{
\item \code{x} is the input vector
\item \code{k} is the exponent
}
}
\details{
If \code{x} is negative, the result is multiplied by -1.
}
\seealso{
Other transform: 
\code{\link{log_trans}()},
\code{\link{symlog_trans}()}
}
\concept{transform}
