% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdeck-package.R
\docType{package}
\name{rdeck-package}
\alias{rdeck-package}
\alias{_PACKAGE}
\title{rdeck: Deck.gl Widget}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Deck.gl widget for R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://qfes.github.io/rdeck}
  \item \url{https://github.com/qfes/rdeck}
  \item Report bugs at \url{https://github.com/qfes/rdeck/issues}
}

}
\author{
\strong{Maintainer}: Anthony North \email{anthony.jl.north@gmail.com}

}
\keyword{internal}
