% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdeck.R, R/widget.R
\docType{package}
\name{rdeck}
\alias{rdeck}
\title{rdeck: A deck.gl widget}
\usage{
rdeck(
  mapbox_api_access_token = Sys.getenv("MAPBOX_ACCESS_TOKEN"),
  map_style = "mapbox://styles/mapbox/dark-v10",
  width = NULL,
  height = NULL,
  elementId = NULL,
  ...
)
}
\arguments{
\item{mapbox_api_access_token}{\code{{character}}}

\item{map_style}{\code{{character}}}

\item{width}{\code{{numeric}}}

\item{height}{\code{{numeric}}}

\item{elementId}{\code{{character}}}

\item{...}{additional parameters to pass to deck}
}
\value{
\code{{rdeck}}
}
\description{
rdeck: A deck.gl widget

Create rdeck widget
}
