% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{rdeck_proxy}
\alias{rdeck_proxy}
\title{RDeck proxy}
\usage{
rdeck_proxy(
  id,
  session = shiny::getDefaultReactiveDomain(),
  map_style = NULL,
  theme = NULL,
  initial_bounds = NULL,
  initial_view_state = NULL,
  controller = NULL,
  picking_radius = NULL,
  use_device_pixels = NULL,
  ...
)
}
\arguments{
\item{id}{the map id}

\item{session}{the shiny session}

\item{map_style}{\code{character}
A mapbox style url. \url{https://docs.mapbox.com/api/maps/#mapbox-styles}}

\item{theme}{The widget theme. Either "kepler" or "light".}

\item{initial_bounds}{\code{sf::st_bbox} | \code{sf::sf} | \code{sf::sfc} | \code{sf::sfg}
The initial bounds of the map; overwrites \code{initial_view_state}.}

\item{initial_view_state}{\code{view_state}
The initial view state of the map. See \code{\link[=view_state]{view_state()}} for details.}

\item{controller}{\code{logical}
If \code{NULL} or \code{FALSE}, the map is not interactive.}

\item{picking_radius}{\code{numeric}
Extra pixels around the pointer to include while picking.}

\item{use_device_pixels}{\code{logical} | \code{numeric}
Controls the resolution of drawing buffer used for rendering.}

\item{...}{additional parameters to pass to the \code{Deck}.}
}
\description{
Creates an \code{\link[=rdeck]{rdeck()}} interface for asynchronous updates of a pre-rendered rdeck map
in Shiny apps.
}
\details{
All rdeck props can be updated through the proxy (\code{NULL} values will be discarded),
layers that are added to the proxy (e.g. \code{rdeck_proxy \%>\% add_h3_hexagon_layer()})
will be merged with pre-rendered rdeck layers.

Layers are merged by their \code{id}. Matched layers will be updated in place, new layers
will be appended and hence drawn \emph{on top} of all existing layers. For layer updates, you
may omit the \code{data} prop to avoid re-serialising unchanged data. All other props will
assume their defaults if omitted.
}
\examples{
\dontrun{
library(shiny)
library(dplyr)
library(h3jsr)
library(viridis)

ui <- fillPage(
  rdeckOutput("map", height = "100\%"),
  absolutePanel(
    top = 10, left = 10,
    sliderInput("range", "value", 0, 1, c(0, 1), step = 0.1)
  )
)

h3_data <- tibble(
  hexagon = get_res0() \%>\%
    get_children(res = 3) \%>\%
    unlist() \%>\%
    unique(),
  value = runif(length(hexagon))
)

map <- rdeck() \%>\%
  add_h3_hexagon_layer(
    id = "h3_hexagon",
    name = "hexagons",
    data = h3_data,
    get_fill_color = scale_color_quantize(
      col = value,
      palette = viridis(6, 0.3)
    ),
    pickable = TRUE,
    auto_highlight = TRUE,
    tooltip = c(hexagon, value)
  )

server <- function(input, output, session) {
  output$map <- renderRdeck(map)

  filtered_data <- reactive({
    h3_data \%>\%
      filter(value >= input$range[1] & value <= input$range[2])
  })

  observe({
    rdeck_proxy("map") \%>\%
      add_h3_hexagon_layer(
        id = "h3_hexagon",
        name = "hexagons",
        data = filtered_data(),
        get_fill_color = scale_color_quantize(
          col = value,
          palette = cividis(6, 0.3)
        ),
        pickable = TRUE,
        auto_highlight = TRUE,
        tooltip = c(hexagon, value)
      )
  })
}

app <- shinyApp(ui, server)
}
}
