% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale_center}
\alias{rescale_center}
\title{Rescale center}
\usage{
rescale_center(scale, center = 0)
}
\arguments{
\item{scale}{<\code{scale}> a scale object}

\item{center}{<\code{number}> the center of the scale input}
}
\description{
Re-centers a scale to have a defined center / midpoint. This is the rdeck equivalent of
\code{\link[scales:rescale_mid]{scales::rescale_mid()}}.

Centering an rdeck scale creates a new scale with the output palette or range centered at \code{center}.
This is similar to creating a diverging scale; the key difference is that the output palette or range
remains linear (with respect to the breaks) and is truncated on the side that is closest to \code{center}.
This is useful in creating \emph{difference} layer, where the output palette or range represents distance
from the center.
}
\examples{
# create a sqrt scale that is centered at 0
rescale_center(
  scale_color_power(col, limits = -36:4),
  center = 0
)

# create a discrete symlog scale that is centered at 5
rescale_center(
  scale_color_threshold(col, limits = -100:100, breaks = breaks_symlog()),
  center = 5
)

}
\seealso{
Other scales: 
\code{\link{rescale_diverge}()},
\code{\link{scale_category}()},
\code{\link{scale_identity}()},
\code{\link{scale_linear}()},
\code{\link{scale_log}()},
\code{\link{scale_power}()},
\code{\link{scale_quantile}()},
\code{\link{scale_quantize}()},
\code{\link{scale_symlog}()},
\code{\link{scale_threshold}()}
}
\concept{scales}
