% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale_diverge}
\alias{rescale_diverge}
\title{Rescale diverge}
\usage{
rescale_diverge(scale, center = 0)
}
\arguments{
\item{scale}{<\code{scale}> a scale object}

\item{center}{<\code{number}> the center of the scale input}
}
\description{
Creates a diverging scale with defined center / midpoint. Similar to \code{\link[=rescale_center]{rescale_center()}}, key difference is
the output palette / range is piecewise linear (with respect to breaks) and the entire output range is
always used.
}
\examples{
# create a diverging linear scale at 0
rescale_diverge(
  scale_color_linear(col, limits = -5:10),
  center = 0
)

# create a diverging log scale at 10
rescale_diverge(
  scale_log(col, limits = 1:1000),
  center = 10
)

}
\seealso{
Other scales: 
\code{\link{rescale_center}()},
\code{\link{scale_category}()},
\code{\link{scale_identity}()},
\code{\link{scale_linear}()},
\code{\link{scale_log}()},
\code{\link{scale_power}()},
\code{\link{scale_quantile}()},
\code{\link{scale_quantize}()},
\code{\link{scale_symlog}()},
\code{\link{scale_threshold}()}
}
\concept{scales}
