% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-proptypes.R
\name{scale}
\alias{scale}
\title{Prop Type: Scale}
\description{
Scales transform input data into visual representations, predominantly colours:
\code{fill}, \code{line}, \code{highlight} and sizes: \code{radius}, \code{elevation}, \code{width}, \code{height}.

Transformations are performed in client-side javascript, thus scales have very minimal
file-size overhead; what you pay for is the source data being scaled & the scale descriptor
(palette, breaks, etc), but you don't pay for the scaled output (e.g. the vector of colours).
Many scales can refer to the same column of a dataframe, but the column will only be serialised once.

Data retrieved in the browser dynamically can also be scaled, including \code{\link[=mvt_layer]{mvt_layer()}}. All \code{rdeck}
scales (except \code{\link[=scale_quantile]{scale_quantile()}}) support both R dataframes and data fetched in javascript, with the
following restrictions:
\itemize{
\item column validation isn't performed -- missing column / incorrect type may error in javascript, or
result in the \code{na_color} / \code{na_value} used for all features
\item \code{limits} / \code{levels} must be explictly defined in the scale -- they cannot be learnt from data
\item quantile scales technically work, but they require a copy of the data to compute the quantiles
}
}
\section{Tidy-eval}{
Columns referenced by a scale support \href{https://dplyr.tidyverse.org/articles/programming.html}{tidy-eval}
expressions. The \code{col} parameter for a scale must be a valid argument to \code{\link[rlang:defusing-advanced]{rlang::ensym()}} and follows
the same rules. Use the \link[rlang:topic-inject]{injection operators} to pass the value of a name
from the environment, e.g. \code{!!my_column}.
}

\section{Legend}{
Scales can optionally produce a colour or numeric legend in the client. Numeric legends give
context for which visual field is being scaled and by what column; colour legends additionally
provide a colour scale and \emph{untransformed} ticks.

Legend ticks can be formatted with a labelling function (e.g. \code{\link[scales:label_number]{scales::label_number()}}). This function
can completely replace the input ticks if you wish; the only constraint is the return value is a character
vector of the same length as input \emph{unformatted ticks}.

Legends are \href{https://kepler.gl}{kepler.gl} inspired.
\if{html}{\out{
  <div class="text-center">
    <img src="https://user-images.githubusercontent.com/391385/103058278-7b1d1000-45ed-11eb-86d2-7ab594bb32de.png" />
  </div>
}}
}

\section{Available scales}{
\itemize{
\item \code{\link{scale_linear}}
\item \code{\link{scale_power}}
\item \code{\link{scale_log}}
\item \code{\link{scale_symlog}}
\item \code{\link{scale_identity}}
\item \code{\link{scale_threshold}}
\item \code{\link{scale_quantile}}
\item \code{\link{scale_category}}
\item \code{\link{scale_quantize}}
}
}

\keyword{internal}
