% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_identity}
\alias{scale_identity}
\title{Scale identity}
\usage{
scale_identity(col, na_value = 0, col_label = "{.col}", legend = TRUE)
}
\arguments{
\item{col}{<\code{name} | \code{string}> The name of the column containing data to be scaled.
Must be a valid input to \code{\link[rlang:defusing-advanced]{rlang::ensym()}}; either a named column (non-standard evaluation), a string.
Supports \href{https://dplyr.tidyverse.org/articles/programming.html}{tidy-eval}.}

\item{na_value}{<\code{number}> The output value for \code{NA} input values.}

\item{col_label}{<\code{string} | \code{function}> A template string or a label function for customising the
column name in the legend.
\itemize{
\item if \code{col_label} is a string, \code{{.col}} may be used to represent the \code{col} name
\item if \code{col_label} is a function, the function must take a single argument: the \code{col} name
}}

\item{legend}{<\code{boolean}> Indicate whether the legend should be displayed for this scale.}
}
\description{
Creates an identity scale; a special case of a linear scale, where input is
mapped to itself (input limits = output range). An identity scale is useful
in cases where input data is already expressed in a visual representation
(e.g. a line width) and should be used as-is.
}
\note{
Identity scales are \emph{almost} equivalent to an \code{\link{accessor}} to a numeric column;
differences are:
\itemize{
\item \code{NA} is replaced with \code{na_value}
\item May render a \emph{numeric} legend
}
}
\seealso{
Other scales: 
\code{\link{scale_category}()},
\code{\link{scale_linear}()},
\code{\link{scale_log}()},
\code{\link{scale_power}()},
\code{\link{scale_quantile}()},
\code{\link{scale_quantize}()},
\code{\link{scale_symlog}()},
\code{\link{scale_threshold}()}
}
\concept{scales}
