% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_linear}
\alias{scale_linear}
\title{Linear Scale}
\usage{
scale_linear(domain = NULL, range = c(0, 1), value, legend = TRUE)
}
\description{
Linear Scale
}
\seealso{
Other scales: 
\code{\link{scale_log}()},
\code{\link{scale_power}()},
\code{\link{scale_quantile}()},
\code{\link{scale_quantize}()},
\code{\link{scale}()}
}
\concept{scales}
