% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_log}
\alias{scale_log}
\alias{scale_color_log}
\title{Scale log}
\usage{
scale_color_log(
  col,
  palette,
  na_color = "#000000",
  base = 10,
  limits = NULL,
  breaks = NULL,
  n_ticks = 6,
  tick_format = format_number,
  legend = TRUE
)

scale_log(
  col,
  range = 0:1,
  na_value = 0,
  base = 10,
  limits = NULL,
  breaks = NULL,
  legend = TRUE
)
}
\arguments{
\item{col}{<\code{name} | \code{string}> The name of the column containing data to be scaled.
Can be either a named column (non-standard evaluation), or an expression evaluating a string.
Supports \href{https://dplyr.tidyverse.org/articles/programming.html}{tidy-eval}.}

\item{palette}{<\code{\link{color}}> The colour palette of the colour scale. Must be a vector of
rgb\link{a} hex strings of length >= 2.}

\item{na_color}{<\code{\link{color}}> The colour value for \code{NA} input values.}

\item{base}{<\code{number}> The log base.}

\item{limits}{<\code{c(min, max)}> The limits of the scale's input. If not null, must be a
numeric vector of length 2, representing \code{c(min, max)}. Values outside the range of
\code{limits} will be clamped to the limits of the scale.}

\item{breaks}{<\code{numeric}> The breaks of the scale, allowing to define a piecewise scale.
If not null, must be \code{length(palette) - 2} or \code{length(range) - 2}, such that each \code{break}
is mapped to a \code{palette} or \code{range} entry.}

\item{n_ticks}{<\code{number}> The number of ticks to display on the legend. Includes the domain
of the scale.}

\item{tick_format}{<\code{function}> A function taking a vector of ticks returning formatted ticks.}

\item{legend}{<\code{logical}> Indicate whether the legend should be displayed for this scale.}

\item{range}{<\code{numeric}> The output range of the numeric scale. Must be be a numeric vector of
length >= 2.}

\item{na_value}{<\code{number}> The output value for \code{NA} input values.}
}
\description{
Creates a continuous log scale. Similar to a \code{\link[=scale_linear]{scale_linear()}} except that an logarithmic
transform is applied to each input.
}
\note{
undefined behaviour if \code{limits} crosses 0.
}
\seealso{
Other scales: 
\code{\link{scale_category}()},
\code{\link{scale_linear}()},
\code{\link{scale_power}()},
\code{\link{scale_quantile}()},
\code{\link{scale_quantize}()},
\code{\link{scale_threshold}()}
}
\concept{scales}
