% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_power}
\alias{scale_power}
\title{Power Scale}
\usage{
scale_power(domain = NULL, range = c(0, 1), exponent = 1, value, legend = TRUE)
}
\description{
Power Scale
}
\seealso{
Other scales: 
\code{\link{scale_linear}()},
\code{\link{scale_log}()},
\code{\link{scale_quantile}()},
\code{\link{scale_quantize}()},
\code{\link{scale}()}
}
\concept{scales}
