% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_threshold}
\alias{scale_threshold}
\alias{scale_color_threshold}
\title{Scale threshold}
\usage{
scale_color_threshold(
  col,
  palette,
  na_color = "#000000",
  limits = NULL,
  breaks = 0.5,
  tick_format = format_number,
  legend = TRUE
)

scale_threshold(
  col,
  range = 0:1,
  na_value = 0,
  limits = NULL,
  breaks = 0.5,
  tick_format = format_number,
  legend = FALSE
)
}
\arguments{
\item{col}{The name of the column containing data to be scaled. Can be either
a named column (non-standard evaluation), or an expression evaluating a string.}

\item{palette}{The color palette of the color scale. Must be a vector of
rgb\link{a} hex strings of length >= 2.}

\item{na_color}{The color value for \code{NA} input values.}

\item{limits}{The limits of the scale's input. If not null, must be a numeric vector
of length 2, representing \code{c(min, max)}. Values outside the range of \code{limits} will be
constrained to the limits of the scale.}

\item{breaks}{The threshold breaks of the scale. Must be \code{length(palette) - 1}
or \code{length(range) - 1}, such that each \code{break} defines boundary between between a
pair of a \code{palette} or \code{range} entries.}

\item{tick_format}{A function taking a vector of ticks returning formatted ticks.}

\item{legend}{Indicate whether the legend should be displayed for this scale.}

\item{range}{The output range of the numeric scale. Must be be a numeric vector of
length >= 2.}

\item{na_value}{The output value for \code{NA} input values.}
}
\description{
Creates a discrete threshold scale. Threshold scale is similar to \code{scale_quantize()}, except
that threshold values can be arbitrary.
}
\seealso{
Other scales: 
\code{\link{scale_category}()},
\code{\link{scale_linear}()},
\code{\link{scale_log}()},
\code{\link{scale_power}()},
\code{\link{scale_quantile}()},
\code{\link{scale_quantize}()}
}
\concept{scales}
