% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenegraph_layer.R
\name{scenegraph_layer}
\alias{scenegraph_layer}
\title{\href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/scenegraph-layer.md}{ScenegraphLayer} deck.gl layer.}
\usage{
scenegraph_layer(
  id = NULL,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  scenegraph = NULL,
  get_scene = NULL,
  get_animator = NULL,
  size_scale = 1,
  size_min_pixels = 0,
  size_max_pixels = 9007199254740991,
  get_position = NULL,
  get_color = NULL,
  get_orientation = NULL,
  get_scale = NULL,
  get_translation = NULL,
  get_transform_matrix = NULL,
  ...
)
}
\arguments{
\item{id}{\code{\link{character}}
The id of the layer. Layer ids must be unique per layer \code{type} for deck.gl
to properly distinguish between them.}

\item{data}{\code{\link{data.frame}} | \code{\link[sf:sf]{sf::sf}}}

\item{visible}{\code{\link{logical}}}

\item{pickable}{\code{\link{logical}}}

\item{opacity}{\code{\link{numeric}}}

\item{position_format}{\code{XY} | \code{XYZ}}

\item{color_format}{\code{RGB} | \code{RGBA}}

\item{auto_highlight}{\code{\link{logical}}}

\item{highlight_color}{\code{\link{integer}}}

\item{scenegraph}{\code{\link{list}}}

\item{get_scene}{\code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{get_animator}{\code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{size_scale}{\code{\link{numeric}}}

\item{size_min_pixels}{\code{\link{numeric}}}

\item{size_max_pixels}{\code{\link{numeric}}}

\item{get_position}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{get_color}{accessor}

\item{get_orientation}{accessor}

\item{get_scale}{accessor}

\item{get_translation}{accessor}

\item{get_transform_matrix}{accessor}

\item{...}{additional layer parameters to pass to deck.gl.
\code{snake_case} parameters will be converted to \code{snakeCase}.}
}
\value{
\code{ScenegraphLayer} & \code{\link{layer}}
A \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/scenegraph-layer.md}{ScenegraphLayer} layer.
Add to an \link{rdeck} map via \code{\link{add_layer}} or \code{\link{rdeck}}.
}
\description{
\href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/scenegraph-layer.md}{ScenegraphLayer} deck.gl layer.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/scenegraph-layer.md}
}
