% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{set_layer_visibility}
\alias{set_layer_visibility}
\title{Set layer visibility}
\usage{
set_layer_visibility(rdeck, id, visible = NULL, visibility_toggle = NULL)
}
\arguments{
\item{rdeck}{<\code{rdeck_proxy}> the rdeck proxy object}

\item{id}{<\code{string}> The layer's identifier must be unique for among all layers of the same
type for a map. Defaults to \code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}, but should be explicitly defined for
updatable layers in a shiny application.}

\item{visible}{<\code{boolean}> Determines whether the layer is visible or not; also determines
whether any legend elements for the layer will be displayed.
When \code{visble = NULL}:
\itemize{
\item on update, use current state in the browser
\item on add, layer is visible
}}

\item{visibility_toggle}{<\code{boolean}> Whether this layer will appear in the layer selector.}
}
\description{
Sets a layer's visibility and whether it is \emph{selectable} in the layer selector.
Setting either \code{visible} or \code{visibility_toggle} as \code{NULL} will have no change
in the browser.
}
