% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{sf_column}
\alias{sf_column}
\title{SF Column}
\usage{
sf_column()
}
\description{
Defines a flag that indicates the active geometry column of an sf object
should be used in a layer's geometry \code{\link[=accessor]{accessor()}}.
}
