% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{sfc_point}
\alias{sfc_point}
\title{SFC Point}
\usage{
sfc_point(x, y, z = NULL, crs = 4326)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{z}{z coordinate {optional}}

\item{crs}{coordinate reference system}
}
\description{
Create an sfc_point column from coordinate vectors
}
