% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny-events}
\alias{shiny-events}
\alias{get_view_bounds}
\alias{get_view_state}
\alias{get_clicked_coordinates}
\alias{get_clicked_layer}
\alias{get_clicked_object}
\alias{get_edited_features}
\title{Shiny events data}
\usage{
get_view_bounds(rdeck, session = shiny::getDefaultReactiveDomain())

get_view_state(rdeck, session = shiny::getDefaultReactiveDomain())

get_clicked_coordinates(rdeck, session = shiny::getDefaultReactiveDomain())

get_clicked_layer(rdeck, session = shiny::getDefaultReactiveDomain())

get_clicked_object(rdeck, session = shiny::getDefaultReactiveDomain())

get_edited_features(rdeck, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{rdeck}{<\code{rdeck_proxy} | \code{string}> the map, or map id}

\item{session}{<\code{ShinySession}> the shiny session}
}
\description{
Utilities for retrieving map event data in a shiny application.
}
\section{Functions}{
\itemize{
\item \code{get_view_bounds()}: Get the current map bounding box

\item \code{get_view_state()}: Get the map view state

\item \code{get_clicked_coordinates()}: Get the last clicked coordinates

\item \code{get_clicked_layer()}: Get the last clicked layer (or NULL)

\item \code{get_clicked_object()}: Get the last clicked object (or NULL)

\item \code{get_edited_features()}: Get the edited features

}}
