% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solid_polygon_layer.R
\name{solid_polygon_layer}
\alias{solid_polygon_layer}
\title{\href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/solid-polygon-layer.md}{SolidPolygonLayer} deck.gl layer.}
\usage{
solid_polygon_layer(
  id = NULL,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  filled = TRUE,
  extruded = FALSE,
  wireframe = FALSE,
  elevation_scale = 1,
  get_polygon = NULL,
  get_elevation = NULL,
  get_fill_color = NULL,
  get_line_color = NULL,
  material = TRUE,
  ...
)
}
\arguments{
\item{id}{\code{\link{character}}
The id of the layer. Layer ids must be unique per layer \code{type} for deck.gl
to properly distinguish between them.}

\item{data}{\code{\link{data.frame}} | \code{\link[sf:sf]{sf::sf}}}

\item{visible}{\code{\link{logical}}}

\item{pickable}{\code{\link{logical}}}

\item{opacity}{\code{\link{numeric}}}

\item{position_format}{\code{XY} | \code{XYZ}}

\item{color_format}{\code{RGB} | \code{RGBA}}

\item{auto_highlight}{\code{\link{logical}}}

\item{highlight_color}{\code{\link{integer}}}

\item{filled}{\code{\link{logical}}}

\item{extruded}{\code{\link{logical}}}

\item{wireframe}{\code{\link{logical}}}

\item{elevation_scale}{\code{\link{numeric}}}

\item{get_polygon}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{get_elevation}{accessor | \code{\link{numeric}}}

\item{get_fill_color}{accessor}

\item{get_line_color}{accessor}

\item{material}{\code{\link{logical}}}

\item{...}{additional layer parameters to pass to deck.gl.
\code{snake_case} parameters will be converted to \code{camelCase}.}
}
\value{
\code{SolidPolygonLayer} & \code{\link{layer}}
A \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/solid-polygon-layer.md}{SolidPolygonLayer} layer.
Add to an \link{rdeck} map via \code{\link{add_layer}} or \code{\link{rdeck}}.
}
\description{
\href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/solid-polygon-layer.md}{SolidPolygonLayer} deck.gl layer.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/solid-polygon-layer.md}
}
