% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{split_geometry}
\alias{split_geometry}
\title{Split multi-geometries for deck.gl compatibility.}
\usage{
split_geometry(obj)
}
\arguments{
\item{obj}{\code{\link[sf:sf]{sf::sf}}
An \link[sf:sf]{sf::sf} object}
}
\value{
\code{\link[sf:sf]{sf::sf}}
}
\description{
Split multi-geometries for deck.gl compatibility.
}
