% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_trans.R
\name{symlog_trans}
\alias{symlog_trans}
\title{SymLog transformation}
\usage{
symlog_trans()
}
\description{
Applies a log1p transform: \code{y = log1p(abs(x))}, where:
\itemize{
\item \code{x} is the input vector
}
}
\details{
If \code{x} is negative, the result is multiplied by -1.
}
\seealso{
Other transform: 
\code{\link{log_trans}()},
\code{\link{power_trans}()}
}
\concept{transform}
