% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deckgl-layers-doc.R, R/deckgl-layers.R
\name{tile_3d_layer}
\alias{tile_3d_layer}
\alias{add_tile_3d_layer}
\title{Tile 3D Layer}
\usage{
add_tile_3d_layer(
  rdeck,
  ...,
  id = NULL,
  name = "Tile3DLayer",
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  get_point_color = "#000000",
  point_size = 1,
  load_options = list(),
  loader = list(id = "3d-tiles", name = "3D Tiles", version = "2.3.6", extensions =
    c("cmpt", "pnts", "b3dm", "i3dm"), mimeTypes = c("application/octet-stream"), tests =
    c("cmpt", "pnts", "b3dm", "i3dm"), parse = NULL, options = list(`3d-tiles` =
    list(loadGLTF = TRUE, decodeQuantizedPositions = FALSE, isTileset = "auto", tile =
    NULL, tileset = NULL, assetGltfUpAxis = NULL))),
  tooltip = FALSE
)
}
\arguments{
\item{id}{\code{string} The layer's identifier must be unique for among all layers of the same
type for a map. Defaults to \code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}, but should be explicitly defined for
updatable layers in a shiny application.}

\item{name}{\code{string} The layer's name identifies the layer on tooltips and legends. It does
not need to be unique, but should be brief. Defaults to the deck.gl class name for the layer.}

\item{data}{data}

\item{visible}{boolean}

\item{pickable}{boolean}

\item{opacity}{number}

\item{position_format}{unknown}

\item{color_format}{unknown}

\item{auto_highlight}{boolean}

\item{highlight_color}{<\code{\link{accessor}}>}

\item{get_point_color}{array}

\item{point_size}{number}

\item{load_options}{object}

\item{loader}{object}

\item{tooltip}{<\code{\link{tooltip}}> Defines the columns (and their order) that will be displayed in
the layer tooltip, if \code{pickable = TRUE}.}
}
\description{
Tile 3D Layer
}
\seealso{
\url{https://github.com/visgl/deck.gl/blob/v8.3.11/docs/api-reference/geo-layers/tile3d-layer.md}

Other layers: 
\code{\link{arc_layer}},
\code{\link{bitmap_layer}},
\code{\link{column_layer}},
\code{\link{contour_layer}},
\code{\link{cpu_grid_layer}},
\code{\link{geojson_layer}},
\code{\link{gpu_grid_layer}},
\code{\link{great_circle_layer}},
\code{\link{grid_cell_layer}},
\code{\link{grid_layer}},
\code{\link{h3_cluster_layer}},
\code{\link{h3_hexagon_layer}},
\code{\link{heatmap_layer}},
\code{\link{hexagon_layer}},
\code{\link{icon_layer}},
\code{\link{line_layer}},
\code{\link{mvt_layer}},
\code{\link{path_layer}},
\code{\link{point_cloud_layer}},
\code{\link{polygon_layer}},
\code{\link{s2_layer}},
\code{\link{scatterplot_layer}},
\code{\link{screen_grid_layer}},
\code{\link{simple_mesh_layer}},
\code{\link{solid_polygon_layer}},
\code{\link{terrain_layer}},
\code{\link{text_layer}},
\code{\link{tile_layer}},
\code{\link{trips_layer}}

Other geo-layers: 
\code{\link{great_circle_layer}},
\code{\link{h3_cluster_layer}},
\code{\link{h3_hexagon_layer}},
\code{\link{mvt_layer}},
\code{\link{s2_layer}},
\code{\link{terrain_layer}},
\code{\link{tile_layer}},
\code{\link{trips_layer}}
}
\concept{geo-layers}
\concept{layers}
