% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_json.R
\name{tile_json}
\alias{tile_json}
\title{Tile json}
\usage{
tile_json(tileset_id, tile_service = NULL)
}
\arguments{
\item{tileset_id}{<\code{string}> The tileset_id, may be one of the following
\itemize{
\item a tileset identifier, e.g. "mapbox.mapbox-streets-v8"
\item a tileset identifier.json, e.g. "mapbox.mapbox-streets-v8.json"
\item a url, e.g. "mapbox://mapbox.mapbox-streets-v8", or "https://mytileserver/tileset"
}

\code{tile_service} is unused if a url is supplied.}

\item{tile_service}{<\code{string}> The tile service name, defaults to
`getOption("rdeck.tile_service") \%||\% "mapbox".}
}
\description{
Intended for use as the \code{data} parameter of \code{\link[=mvt_layer]{mvt_layer()}}. Creates a tile json url
from a tileset_id and tile_service. Tile service defaults to
\verb{getOption("rdeck.tile_service") \%||\% "mapbox}.

The created url will be fetched and parsed with \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}.
}
\note{
Authentication via \code{\link[=mapbox_access_token]{mapbox_access_token()}} occurs when \code{tile_service = "mapbox"},
or \code{tileset_id} uses the mapbox scheme (i.e. mapbox://).
}
\examples{
tile_json("mapbox.mapbox-streets-v8", "mapbox")
tile_json("mapbox.mapbox-streets-v8.json", "mapbox")
tile_json("mapbox://mapbox.mapbox-streets-v8")
tile_json("mapbox://mapbox.mapbox-streets-v8.json")
tile_json("https://mytileserver/tileset.json")
tile_json("tileset.json", "https://mytileserver")

}
