% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deckgl-layers-doc.R, R/deckgl-layers.R
\name{tile_layer}
\alias{tile_layer}
\alias{add_tile_layer}
\title{Tile Layer}
\usage{
add_tile_layer(
  rdeck,
  ...,
  id = NULL,
  name = "TileLayer",
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  get_tile_data = NULL,
  extent = NULL,
  tile_size = 512,
  max_zoom = NULL,
  min_zoom = 0,
  max_cache_size = NULL,
  max_cache_byte_size = NULL,
  refinement_strategy = "best-available",
  z_range = NULL,
  max_requests = 6,
  tooltip = FALSE
)
}
\description{
Tile Layer
}
\seealso{
\url{https://github.com/visgl/deck.gl/blob/v8.3.11/docs/api-reference/geo-layers/tile-layer.md}

Other layers: 
\code{\link{arc_layer}},
\code{\link{bitmap_layer}},
\code{\link{column_layer}},
\code{\link{contour_layer}},
\code{\link{cpu_grid_layer}},
\code{\link{geojson_layer}},
\code{\link{gpu_grid_layer}},
\code{\link{great_circle_layer}},
\code{\link{grid_cell_layer}},
\code{\link{grid_layer}},
\code{\link{h3_cluster_layer}},
\code{\link{h3_hexagon_layer}},
\code{\link{heatmap_layer}},
\code{\link{hexagon_layer}},
\code{\link{icon_layer}},
\code{\link{line_layer}},
\code{\link{mvt_layer}},
\code{\link{path_layer}},
\code{\link{point_cloud_layer}},
\code{\link{polygon_layer}},
\code{\link{s2_layer}},
\code{\link{scatterplot_layer}},
\code{\link{screen_grid_layer}},
\code{\link{simple_mesh_layer}},
\code{\link{solid_polygon_layer}},
\code{\link{terrain_layer}},
\code{\link{text_layer}},
\code{\link{tile_3d_layer}},
\code{\link{trips_layer}}

Other geo-layers: 
\code{\link{great_circle_layer}},
\code{\link{h3_cluster_layer}},
\code{\link{h3_hexagon_layer}},
\code{\link{mvt_layer}},
\code{\link{s2_layer}},
\code{\link{terrain_layer}},
\code{\link{tile_3d_layer}},
\code{\link{trips_layer}}
}
\concept{geo-layers}
\concept{layers}
