% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{tooltip}
\alias{tooltip}
\title{Tooltip}
\usage{
tooltip(expr)
}
\arguments{
\item{expr}{The expr to evaluate / deparse.
\itemize{
\item Literals and calls other than \code{c()} are evaluated
\item Names are deparsed
\item Arguments to \code{c()} are modified using the logic of the above.
}}
}
\value{
evaluated expression.
}
\description{
Create a character vector representing names to be included in a tooltip
}
\keyword{internal}
