% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deckgl-layers-doc.R, R/deckgl-layers.R
\name{trips_layer}
\alias{trips_layer}
\alias{add_trips_layer}
\title{Trips Layer}
\usage{
add_trips_layer(
  rdeck,
  ...,
  id = NULL,
  name = "TripsLayer",
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  width_units = "meters",
  width_scale = 1,
  width_min_pixels = 0,
  width_max_pixels = 9007199254740991,
  rounded = FALSE,
  miter_limit = 4,
  billboard = FALSE,
  get_path = path,
  get_color = "#000000ff",
  get_width = 1,
  trail_length = 120,
  current_time = 0,
  get_timestamps = NULL,
  tooltip = FALSE
)
}
\arguments{
\item{id}{\code{string} The layer's identifier must be unique for among all layers of the same
type for a map. Defaults to \code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}, but should be explicitly defined for
updatable layers in a shiny application.}

\item{name}{\code{string} The layer's name identifies the layer on tooltips and legends. It does
not need to be unique, but should be brief. Defaults to the deck.gl class name for the layer.}

\item{data}{data}

\item{visible}{boolean}

\item{pickable}{boolean}

\item{opacity}{number}

\item{position_format}{unknown}

\item{color_format}{unknown}

\item{auto_highlight}{boolean}

\item{highlight_color}{<\code{\link{accessor}}>}

\item{width_units}{unknown}

\item{width_scale}{number}

\item{width_min_pixels}{number}

\item{width_max_pixels}{number}

\item{rounded}{boolean}

\item{miter_limit}{number}

\item{billboard}{boolean}

\item{get_path}{<\code{\link{accessor}}>}

\item{get_color}{<\code{\link{accessor}}>}

\item{get_width}{<\code{\link{accessor}}>}

\item{trail_length}{number}

\item{current_time}{number}

\item{get_timestamps}{<\code{\link{accessor}}>}

\item{tooltip}{<\code{\link{tooltip}}> Defines the columns (and their order) that will be displayed in
the layer tooltip, if \code{pickable = TRUE}.}
}
\description{
Trips Layer
}
\seealso{
\url{https://github.com/visgl/deck.gl/blob/v8.3.11/docs/api-reference/geo-layers/trips-layer.md}

Other layers: 
\code{\link{arc_layer}},
\code{\link{bitmap_layer}},
\code{\link{column_layer}},
\code{\link{contour_layer}},
\code{\link{cpu_grid_layer}},
\code{\link{geojson_layer}},
\code{\link{gpu_grid_layer}},
\code{\link{great_circle_layer}},
\code{\link{grid_cell_layer}},
\code{\link{grid_layer}},
\code{\link{h3_cluster_layer}},
\code{\link{h3_hexagon_layer}},
\code{\link{heatmap_layer}},
\code{\link{hexagon_layer}},
\code{\link{icon_layer}},
\code{\link{line_layer}},
\code{\link{mvt_layer}},
\code{\link{path_layer}},
\code{\link{point_cloud_layer}},
\code{\link{polygon_layer}},
\code{\link{s2_layer}},
\code{\link{scatterplot_layer}},
\code{\link{screen_grid_layer}},
\code{\link{simple_mesh_layer}},
\code{\link{solid_polygon_layer}},
\code{\link{terrain_layer}},
\code{\link{text_layer}},
\code{\link{tile_3d_layer}},
\code{\link{tile_layer}}

Other geo-layers: 
\code{\link{great_circle_layer}},
\code{\link{h3_cluster_layer}},
\code{\link{h3_hexagon_layer}},
\code{\link{mvt_layer}},
\code{\link{s2_layer}},
\code{\link{terrain_layer}},
\code{\link{tile_3d_layer}},
\code{\link{tile_layer}}
}
\concept{geo-layers}
\concept{layers}
