% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{visit}
\alias{visit}
\title{Simple expression visitor for translating R into accessors}
\usage{
visit(expr, names, columnar = TRUE)
}
\arguments{
\item{expr}{\code{\link{name}} | \code{\link{call}} | value
The expression to visit.}

\item{names}{\code{\link{character}}
Vector of names to expand into either \code{data.frame[expr][index]}
or \code{data[expr]}, depending on the value of \code{columnar} parameter}

\item{columnar}{\code{\link{logical}}
Is the data passed to layers a columnar table, or an array of objects?}
}
\value{
\code{\link[htmlwidgets:JS]{htmlwidgets::JS}} | \code{\link{name}} | \code{\link{call}} | value
Either a \link[htmlwidgets:JS]{htmlwidgets::JS} instance or the original expression.
}
\description{
Simple expression visitor for translating R into accessors
}
\keyword{internal}
