% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-proptypes.R
\name{wk-geometry}
\alias{wk-geometry}
\title{Prop Type: WK Geometry}
\description{
Geometry vectors are parsed with \code{\link[wk:wk_handle]{wk::wk_handle()}}. All geometry formats
that \{wk\} supports directly are available to use in {rdeck} layers. You may
also use geometry formats which \{wk\} doesn't directly support, provided a
\emph{loaded} package defines a \{wk\} \emph{reader} (i.e. \code{\link[wk:wk_handle]{wk::wk_handle()}} S3 method)
for the geometry type.

Geometry formats supported by \{wk\} (always supported):
\itemize{
\item \code{\link[wk:wkb]{wk::wkb()}}
\item \code{\link[wk:wkt]{wk::wkt()}}
\item \code{\link[wk:xy]{wk::xy()}} (\code{m} dimension is always ignored)
\item \code{\link[wk:crc]{wk::crc()}}
\item \code{\link[wk:grd]{wk::grd()}}
\item \code{\link[wk:rct]{wk::rct()}}
\item \code{\link[sf:sfc]{sf::st_sfc()}}
}

Geometry formats supported by other packages (not dependencies of \{rdeck\}):
\itemize{
\item \code{\link[s2:as_s2_geography]{s2::s2_geography()}}
\item \code{\link[geos:as_geos_geometry]{geos::geos_geometry()}}
\item Others?
}
}
\keyword{internal}
