% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{xy}
\alias{xy}
\alias{xyz}
\alias{sfc_point}
\title{Point vectors}
\usage{
xy(x = double(), y = double(), crs = wk::wk_crs_longlat())

xyz(x = double(), y = double(), z = double(), crs = wk::wk_crs_longlat())

sfc_point(x = double(), y = double(), z = NULL, crs = wk::wk_crs_longlat())
}
\arguments{
\item{crs}{A value to be propagated as the CRS for this vector.}
}
\description{
Create point vectors
}
\section{Functions}{
\itemize{
\item \code{xy()}: Efficient 2-dimensional point vector

\item \code{xyz()}: Efficient 3-dimensional point vector

\item \code{sfc_point()}: Simple features point vector

}}
\examples{
xy(1:5, 1:5)
xyz(1:5, 1:5, 1:5)
sfc_point(1:5, 1:5)
sfc_point(1:5, 1:5, 1:5)
}
