% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{accessor}
\alias{accessor}
\title{Create accessors for deck.gl layers.}
\usage{
accessor(data, expr)
}
\arguments{
\item{data}{\code{{data.frame}}
used as expr environment}

\item{expr}{\code{{name | call | atomic}}
an expression typically created with \code{\link[base:substitute]{base::substitute()}}}
}
\value{
\code{{JS | expr}}
}
\description{
Create accessors for deck.gl layers.
}
