% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{accessor}
\alias{accessor}
\title{Create accessors for deck.gl layers.}
\usage{
accessor(expr, data = NULL, columnar = TRUE)
}
\arguments{
\item{expr}{\code{\link{name}} | \code{\link{call}} | value
An expression or a value.}

\item{data}{\code{\link{data.frame}}
Each \link{name} in \code{expr} that appears in \code{names(data)} will be converted into
one of the following:
\itemize{
\item \code{data.frame["name"][index]} for \code{columnar} == \link{TRUE}
\item \code{data["name"]} else
}}

\item{columnar}{\code{\link{logical}}
Will the layer data in javascript be a columnar table (an object with array
properties), or an array of objects?}
}
\value{
[\verb{htmlwidgets::JS]} | \code{eval(expr)}
Either a \link[htmlwidgets:JS]{htmlwidgets::JS} or evaluated \code{expr}
}
\description{
Create accessors for deck.gl layers.
}
\keyword{internal}
