% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{accessor}
\alias{accessor}
\title{Create accessors for deck.gl layers.}
\usage{
accessor(expr, data = NULL, columnar = TRUE)
}
\arguments{
\item{expr}{\code{{name | call | atomic}}
an expression typically created with \code{\link[base:substitute]{base::substitute()}}}

\item{data}{\code{{data.frame}}
used as expr environment}

\item{columnar}{\code{{logical}}
Is the data passed to layers a columnar table, or an array of objects?}
}
\value{
\code{{JS | expr}}
Either a \link[htmlwidgets:JS]{htmlwidgets::JS} or \code{eval(expr)}
}
\description{
Create accessors for deck.gl layers.
}
