% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitmap_layer.R
\name{add_bitmap_layer}
\alias{add_bitmap_layer}
\title{Add a \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/bitmap-layer.md}{BitmapLayer} deck.gl layer to an \link{rdeck} map.}
\usage{
add_bitmap_layer(
  rdeck,
  id = NULL,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  image = NULL,
  bounds = c(1, 0, 0, 1),
  desaturate = 0,
  transparent_color = c(0, 0, 0, 0),
  tint_color = c(255, 255, 255),
  ...
)
}
\arguments{
\item{rdeck}{\code{\link{rdeck}}
An \link{rdeck} map.}

\item{id}{\code{\link{character}}
The id of the layer. Layer ids must be unique per layer \code{type} for deck.gl
to properly distinguish between them.}

\item{data}{\code{\link{data.frame}} | \code{\link[sf:sf]{sf::sf}}}

\item{visible}{\code{\link{logical}}}

\item{pickable}{\code{\link{logical}}}

\item{opacity}{\code{\link{numeric}}}

\item{position_format}{\code{XY} | \code{XYZ}}

\item{color_format}{\code{RGB} | \code{RGBA}}

\item{auto_highlight}{\code{\link{logical}}}

\item{highlight_color}{\code{\link{integer}}}

\item{image}{\code{\link{image}} | \code{\link{character}}}

\item{bounds}{\code{\link{numeric}}}

\item{desaturate}{\code{\link{numeric}}}

\item{transparent_color}{\code{\link{integer}}}

\item{tint_color}{\code{\link{integer}}}

\item{...}{
  Arguments passed on to \code{\link[=bitmap_layer]{bitmap_layer}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\code{\link{rdeck}}
The \link{rdeck} map.
}
\description{
Add a \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/bitmap-layer.md}{BitmapLayer} deck.gl layer to an \link{rdeck} map.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/bitmap-layer.md}
}
