% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_layer.R
\name{add_grid_layer}
\alias{add_grid_layer}
\title{Add a \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/grid-layer.md}{GridLayer} deck.gl layer to an \link{rdeck} map.}
\usage{
add_grid_layer(
  rdeck,
  id = NULL,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  color_domain = NULL,
  color_range = list(c(255, 255, 178), c(254, 217, 118), c(254, 178, 76), c(253, 141,
    60), c(240, 59, 32), c(189, 0, 38)),
  get_color_weight = NULL,
  color_aggregation = "SUM",
  elevation_domain = NULL,
  elevation_range = c(0, 1000),
  get_elevation_weight = NULL,
  elevation_aggregation = "SUM",
  elevation_scale = 1,
  cell_size = 1000,
  coverage = 1,
  get_position = NULL,
  extruded = FALSE,
  material = TRUE,
  get_color_value = NULL,
  lower_percentile = 0,
  upper_percentile = 100,
  color_scale_type = "quantize",
  get_elevation_value = NULL,
  elevation_lower_percentile = 0,
  elevation_upper_percentile = 100,
  elevation_scale_type = "linear",
  grid_aggregator = NULL,
  gpu_aggregation = FALSE,
  ...
)
}
\arguments{
\item{rdeck}{\code{\link{rdeck}}
An \link{rdeck} map.}

\item{id}{\code{\link{character}}
The id of the layer. Layer ids must be unique per layer \code{type} for deck.gl
to properly distinguish between them.}

\item{data}{\code{\link{data.frame}} | \code{\link[sf:sf]{sf::sf}}}

\item{visible}{\code{\link{logical}}}

\item{pickable}{\code{\link{logical}}}

\item{opacity}{\code{\link{numeric}}}

\item{position_format}{\code{XY} | \code{XYZ}}

\item{color_format}{\code{RGB} | \code{RGBA}}

\item{auto_highlight}{\code{\link{logical}}}

\item{highlight_color}{\code{\link{integer}}}

\item{color_domain}{\code{\link{numeric}}}

\item{color_range}{\code{\link{list}}}

\item{get_color_weight}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{color_aggregation}{\code{SUM} | \code{MEAN} | \code{MIN} | \code{MAX}}

\item{elevation_domain}{\code{\link{numeric}}}

\item{elevation_range}{\code{\link{numeric}}}

\item{get_elevation_weight}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{elevation_aggregation}{\code{SUM} | \code{MEAN} | \code{MIN} | \code{MAX}}

\item{elevation_scale}{\code{\link{numeric}}}

\item{cell_size}{\code{\link{numeric}}}

\item{coverage}{\code{\link{numeric}}}

\item{get_position}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{extruded}{\code{\link{logical}}}

\item{material}{\code{\link{logical}}}

\item{get_color_value}{accessor}

\item{lower_percentile}{\code{\link{numeric}}}

\item{upper_percentile}{\code{\link{numeric}}}

\item{color_scale_type}{\code{quantize} | \code{linear} | \code{quantile} | \code{ordinal}}

\item{get_elevation_value}{accessor}

\item{elevation_lower_percentile}{\code{\link{numeric}}}

\item{elevation_upper_percentile}{\code{\link{numeric}}}

\item{elevation_scale_type}{\code{quantize} | \code{linear} | \code{quantile} | \code{ordinal}}

\item{grid_aggregator}{\code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{gpu_aggregation}{\code{\link{logical}}}

\item{...}{
  Arguments passed on to \code{\link[=grid_layer]{grid_layer}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\code{\link{rdeck}}
The \link{rdeck} map.
}
\description{
Add a \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/grid-layer.md}{GridLayer} deck.gl layer to an \link{rdeck} map.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/grid-layer.md}
}
