% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_heatmap_layer.R
\name{add_heatmap_layer}
\alias{add_heatmap_layer}
\title{Add HeatmapLayer to an rdeck map.}
\usage{
add_heatmap_layer(
  rdeck,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  get_position = NULL,
  get_weight = NULL,
  intensity = 1,
  radius_pixels = 50,
  color_range = list(c(255, 255, 178), c(254, 217, 118), c(254, 178, 76), c(253, 141,
    60), c(240, 59, 32), c(189, 0, 38)),
  threshold = 0.05,
  color_domain = NULL,
  ...
)
}
\arguments{
\item{rdeck}{\\verb{\{rdeck\}\\} an rdeck widget instance}

\item{data}{\code{{data.frame | sf}}}

\item{visible}{\code{{logical}}}

\item{pickable}{\code{{logical}}}

\item{opacity}{\code{{numeric}}}

\item{position_format}{\code{{"XY" | "XYZ"}}}

\item{color_format}{\code{{"RGB" | "RGBA"}}}

\item{auto_highlight}{\code{{logical}}}

\item{highlight_color}{\code{{integer}}}

\item{get_position}{\code{{accessor | JS}}}

\item{get_weight}{\code{{accessor | numeric}}}

\item{intensity}{\code{{numeric}}}

\item{radius_pixels}{\code{{numeric}}}

\item{color_range}{\code{{list}}}

\item{threshold}{\code{{numeric}}}

\item{color_domain}{\code{{numeric}}}

\item{...}{additional layer parameters to pass to deck.gl}
}
\value{
\\verb{\{rdeck\}\\}
}
\description{
Add HeatmapLayer to an rdeck map.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/heatmap-layer.md}
}
