% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_layer.R
\name{add_layer}
\alias{add_layer}
\title{Add a deck.gl \href{https://github.com/uber/deck.gl/blob/master/docs/api-reference/layer.md}{layer} to an rdeck map.}
\usage{
add_layer(rdeck, layer)
}
\arguments{
\item{rdeck}{\code{\link{rdeck}}
An \link{rdeck} map.}

\item{layer}{\code{\link{layer}}}
}
\value{
\code{\link{rdeck}}
The \link{rdeck} map.
}
\description{
Add a deck.gl \href{https://github.com/uber/deck.gl/blob/master/docs/api-reference/layer.md}{layer} to an rdeck map.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/master/docs/api-reference/layer.md}
}
