% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{add_layer}
\alias{add_layer}
\title{Add Layer}
\usage{
add_layer(rdeck, layer)
}
\arguments{
\item{rdeck}{\code{rdeck}
An rdeck map}

\item{layer}{The layer}
}
\description{
Add a deck.gl layer to an rdeck map.
}
\keyword{internal}
