% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_layer.R
\name{add_polygon_layer}
\alias{add_polygon_layer}
\title{PolygonLayer}
\usage{
add_polygon_layer(
  rdeck,
  id = "PolygonLayer",
  data = data.frame(),
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  stroked = TRUE,
  filled = TRUE,
  extruded = FALSE,
  elevation_scale = 1,
  wireframe = FALSE,
  line_width_units = "meters",
  line_width_scale = 1,
  line_width_min_pixels = 0,
  line_width_max_pixels = 9007199254740991,
  line_joint_rounded = FALSE,
  line_miter_limit = 4,
  get_polygon = polygon,
  get_fill_color = c(0, 0, 0, 255),
  get_line_color = c(0, 0, 0, 255),
  get_line_width = 1,
  get_elevation = 1000,
  material = TRUE,
  ...
)
}
\arguments{
\item{rdeck}{\code{rdeck}}

\item{id}{\code{character}
The id of the layer. Layer ids must be unique per layer \code{type} for deck.gl
to properly distinguish between them.}

\item{data}{\code{data.frame} | \code{sf::sf}
Deck.gl layer data, either a \link{data.frame} or \link[sf:sf]{sf::sf}.}

\item{visible}{\code{logical}
Whether the layer is visible.}

\item{pickable}{\code{logical}
Whether the layer responds to mouse pointer picking events.}

\item{opacity}{\code{numeric}
The opacity of the layer.}

\item{position_format}{\code{"XY"} | \code{"XYZ"}
Determines whether each coordinate has two (XY) or three (XYZ) numbers.}

\item{color_format}{\code{"RGB"} | \code{"RGBA"}
RGB will make the layer ignore the alpha channel of colours returned by
accessors. Opacity controlled by \code{opacity} is still applied.}

\item{auto_highlight}{\code{logical}
When \code{TRUE}, current object pointed by mouse pointer (when hovered over) is
highlighted with highlight_color. Requires \code{pickable} to be \code{TRUE}.}

\item{highlight_color}{\code{integer} vector of (R,G,B) or (R,G,B,A)
RGBA color to be used to render highlighted object. When 3 component (RGB)
array is specified, a default value of 255 is used for alpha.}

\item{...}{additional deck.gl layer parameters}
}
\description{
PolygonLayer
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.17/docs/layers/polygon-layer.md}

Other add_layers: 
\code{\link{add_arc_layer}()},
\code{\link{add_bitmap_layer}()},
\code{\link{add_column_layer}()},
\code{\link{add_contour_layer}()},
\code{\link{add_cpu_grid_layer}()},
\code{\link{add_geojson_layer}()},
\code{\link{add_gpu_grid_layer}()},
\code{\link{add_great_circle_layer}()},
\code{\link{add_grid_cell_layer}()},
\code{\link{add_grid_layer}()},
\code{\link{add_h3_cluster_layer}()},
\code{\link{add_h3_hexagon_layer}()},
\code{\link{add_heatmap_layer}()},
\code{\link{add_hexagon_layer}()},
\code{\link{add_icon_layer}()},
\code{\link{add_line_layer}()},
\code{\link{add_path_layer}()},
\code{\link{add_point_cloud_layer}()},
\code{\link{add_s2_layer}()},
\code{\link{add_scatterplot_layer}()},
\code{\link{add_scenegraph_layer}()},
\code{\link{add_screen_grid_layer}()},
\code{\link{add_simple_mesh_layer}()},
\code{\link{add_solid_polygon_layer}()},
\code{\link{add_text_layer}()},
\code{\link{add_tile3d_layer}()},
\code{\link{add_tile_layer}()},
\code{\link{add_trips_layer}()}
}
\concept{add_layers}
