% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_layer.R
\name{add_s2_layer}
\alias{add_s2_layer}
\title{Add a \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/s2-layer.md}{S2Layer} deck.gl layer to an \link{rdeck} map.}
\usage{
add_s2_layer(
  rdeck,
  id = NULL,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  get_s2token = NULL,
  stroked = TRUE,
  filled = TRUE,
  extruded = FALSE,
  elevation_scale = 1,
  wireframe = FALSE,
  line_width_units = "meters",
  line_width_scale = 1,
  line_width_min_pixels = 0,
  line_width_max_pixels = 9007199254740991,
  line_joint_rounded = FALSE,
  line_miter_limit = 4,
  get_polygon = NULL,
  get_fill_color = NULL,
  get_line_color = NULL,
  get_line_width = NULL,
  get_elevation = NULL,
  material = TRUE,
  ...
)
}
\arguments{
\item{rdeck}{\code{\link{rdeck}}
An \link{rdeck} map.}

\item{id}{\code{\link{character}}
The id of the layer. Layer ids must be unique per layer \code{type} for deck.gl
to properly distinguish between them.}

\item{data}{\code{\link{data.frame}} | \code{\link[sf:sf]{sf::sf}}}

\item{visible}{\code{\link{logical}}}

\item{pickable}{\code{\link{logical}}}

\item{opacity}{\code{\link{numeric}}}

\item{position_format}{\code{XY} | \code{XYZ}}

\item{color_format}{\code{RGB} | \code{RGBA}}

\item{auto_highlight}{\code{\link{logical}}}

\item{highlight_color}{\code{\link{integer}}}

\item{get_s2token}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{stroked}{\code{\link{logical}}}

\item{filled}{\code{\link{logical}}}

\item{extruded}{\code{\link{logical}}}

\item{elevation_scale}{\code{\link{numeric}}}

\item{wireframe}{\code{\link{logical}}}

\item{line_width_units}{\code{pixels} | \code{meters}}

\item{line_width_scale}{\code{\link{numeric}}}

\item{line_width_min_pixels}{\code{\link{numeric}}}

\item{line_width_max_pixels}{\code{\link{numeric}}}

\item{line_joint_rounded}{\code{\link{logical}}}

\item{line_miter_limit}{\code{\link{numeric}}}

\item{get_polygon}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{get_fill_color}{accessor}

\item{get_line_color}{accessor}

\item{get_line_width}{accessor | \code{\link{numeric}}}

\item{get_elevation}{accessor | \code{\link{numeric}}}

\item{material}{\code{\link{logical}}}

\item{...}{
  Arguments passed on to \code{\link[=s2_layer]{s2_layer}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\code{\link{rdeck}}
The \link{rdeck} map.
}
\description{
Add a \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/s2-layer.md}{S2Layer} deck.gl layer to an \link{rdeck} map.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/s2-layer.md}
}
