% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_grid_layer.R
\name{add_screen_grid_layer}
\alias{add_screen_grid_layer}
\title{Add a \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/screen-grid-layer.md}{ScreenGridLayer} deck.gl layer to an \link{rdeck} map.}
\usage{
add_screen_grid_layer(
  rdeck,
  id = NULL,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  cell_size_pixels = 100,
  cell_margin_pixels = 2,
  color_domain = NULL,
  color_range = list(c(255, 255, 178), c(254, 217, 118), c(254, 178, 76), c(253, 141,
    60), c(240, 59, 32), c(189, 0, 38)),
  get_position = NULL,
  get_weight = NULL,
  gpu_aggregation = TRUE,
  aggregation = "SUM",
  ...
)
}
\arguments{
\item{rdeck}{\code{\link{rdeck}}
An \link{rdeck} map.}

\item{id}{\code{\link{character}}
The id of the layer. Layer ids must be unique per layer \code{type} for deck.gl
to properly distinguish between them.}

\item{data}{\code{\link{data.frame}} | \code{\link[sf:sf]{sf::sf}}}

\item{visible}{\code{\link{logical}}}

\item{pickable}{\code{\link{logical}}}

\item{opacity}{\code{\link{numeric}}}

\item{position_format}{\code{XY} | \code{XYZ}}

\item{color_format}{\code{RGB} | \code{RGBA}}

\item{auto_highlight}{\code{\link{logical}}}

\item{highlight_color}{\code{\link{integer}}}

\item{cell_size_pixels}{\code{\link{numeric}}}

\item{cell_margin_pixels}{\code{\link{numeric}}}

\item{color_domain}{\code{\link{numeric}}}

\item{color_range}{\code{\link{list}}}

\item{get_position}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{get_weight}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{gpu_aggregation}{\code{\link{logical}}}

\item{aggregation}{\code{SUM} | \code{MEAN} | \code{MIN} | \code{MAX}}

\item{...}{
  Arguments passed on to \code{\link[=screen_grid_layer]{screen_grid_layer}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\code{\link{rdeck}}
The \link{rdeck} map.
}
\description{
Add a \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/screen-grid-layer.md}{ScreenGridLayer} deck.gl layer to an \link{rdeck} map.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/screen-grid-layer.md}
}
