% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tile3d_layer.R
\name{add_tile3d_layer}
\alias{add_tile3d_layer}
\title{Add Tile3DLayer to an rdeck map.}
\usage{
add_tile3d_layer(
  rdeck,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  get_point_color = c(0, 0, 0),
  point_size = 1,
  load_options = NULL,
  ...
)
}
\arguments{
\item{rdeck}{\\verb{\{rdeck\}\\} an rdeck widget instance}

\item{data}{\code{{data.frame | sf}}}

\item{visible}{\code{{logical}}}

\item{pickable}{\code{{logical}}}

\item{opacity}{\code{{numeric}}}

\item{position_format}{\code{{"XY" | "XYZ"}}}

\item{color_format}{\code{{"RGB" | "RGBA"}}}

\item{auto_highlight}{\code{{logical}}}

\item{highlight_color}{\code{{integer}}}

\item{get_point_color}{\code{{numeric}}}

\item{point_size}{\code{{numeric}}}

\item{load_options}{\code{{list}}}

\item{...}{additional layer parameters to pass to deck.gl}
}
\value{
\\verb{\{rdeck\}\\}
}
\description{
Add Tile3DLayer to an rdeck map.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/tile-3d-layer.md}
}
