% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_trips_layer.R
\name{add_trips_layer}
\alias{add_trips_layer}
\title{Add a \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/trips-layer.md}{TripsLayer} deck.gl layer to an \link{rdeck} map.}
\usage{
add_trips_layer(
  rdeck,
  id = NULL,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  width_units = "meters",
  width_scale = 1,
  width_min_pixels = 0,
  width_max_pixels = 9007199254740991,
  rounded = FALSE,
  miter_limit = 4,
  billboard = FALSE,
  get_path = NULL,
  get_color = NULL,
  get_width = NULL,
  trail_length = 120,
  current_time = 0,
  get_timestamps = NULL,
  ...
)
}
\arguments{
\item{rdeck}{\code{\link{rdeck}}
An \link{rdeck} map.}

\item{id}{\code{\link{character}}
The id of the layer. Layer ids must be unique per layer \code{type} for deck.gl
to properly distinguish between them.}

\item{data}{\code{\link{data.frame}} | \code{\link[sf:sf]{sf::sf}}}

\item{visible}{\code{\link{logical}}}

\item{pickable}{\code{\link{logical}}}

\item{opacity}{\code{\link{numeric}}}

\item{position_format}{\code{XY} | \code{XYZ}}

\item{color_format}{\code{RGB} | \code{RGBA}}

\item{auto_highlight}{\code{\link{logical}}}

\item{highlight_color}{\code{\link{integer}}}

\item{width_units}{\code{pixels} | \code{meters}}

\item{width_scale}{\code{\link{numeric}}}

\item{width_min_pixels}{\code{\link{numeric}}}

\item{width_max_pixels}{\code{\link{numeric}}}

\item{rounded}{\code{\link{logical}}}

\item{miter_limit}{\code{\link{numeric}}}

\item{billboard}{\code{\link{logical}}}

\item{get_path}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{get_color}{accessor}

\item{get_width}{accessor | \code{\link{numeric}}}

\item{trail_length}{\code{\link{numeric}}}

\item{current_time}{\code{\link{numeric}}}

\item{get_timestamps}{accessor}

\item{...}{
  Arguments passed on to \code{\link[=trips_layer]{trips_layer}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
\code{\link{rdeck}}
The \link{rdeck} map.
}
\description{
Add a \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/trips-layer.md}{TripsLayer} deck.gl layer to an \link{rdeck} map.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/trips-layer.md}
}
