% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{bounds}
\alias{bounds}
\title{Create a bounds instance from a simple features object, column or geometry}
\usage{
bounds(obj)
}
\arguments{
\item{obj}{\code{\link[sf:sf]{sf::sf}} | \code{\link[sf:sfc]{sf::sfc}} | sf::sfg`
A simple features object, column or geometry to compute bounds from}
}
\value{
\code{bounds}
A vector of \link[sf:st_point]{sf::st_point}
}
\description{
Create a bounds instance from a simple features object, column or geometry
}
