% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{editor_options}
\alias{editor_options}
\title{Polygon editor options}
\usage{
editor_options(mode = cur_value(), features = cur_value())
}
\arguments{
\item{mode}{<\code{editor-mode}> The polygon editor mode. One of:
\itemize{
\item \code{view}: editor is in readonly mode
\item \code{select}: select/unselect features
\item \code{modify}: add/move/delete vertices
\item \code{transform}: move/scale/rotate selected features
\item \code{point}: draw points
\item \code{linestring}: draw linestrings by clicking each vertex
\item \code{polygon}: draw polygons by clicking each vertex
\item \code{lasso}: freehand polygon draw by click-dragging
}}

\item{features}{<\code{\link{wk-geometry}}> Features with which to initialise the editor.
Requires CRS \href{http://epsg.io/4326}{EPSG:4326}.}
}
\description{
Options for the polygon editor
}
