% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deckgl-layers-doc.R, R/deckgl-layers.R
\name{grid_cell_layer}
\alias{grid_cell_layer}
\alias{add_grid_cell_layer}
\title{Grid Cell Layer}
\usage{
add_grid_cell_layer(
  rdeck,
  ...,
  id = NULL,
  name = "GridCellLayer",
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  disk_resolution = 20,
  vertices = NULL,
  radius = 1000,
  angle = 0,
  offset = c(1, 1),
  coverage = 1,
  elevation_scale = 1,
  line_width_units = "meters",
  line_width_scale = 1,
  line_width_min_pixels = 0,
  line_width_max_pixels = 9007199254740991,
  extruded = TRUE,
  wireframe = FALSE,
  filled = TRUE,
  stroked = FALSE,
  get_position = position,
  get_fill_color = "#000000ff",
  get_line_color = "#000000ff",
  get_line_width = 1,
  get_elevation = 1000,
  material = TRUE,
  cell_size = 1000,
  tooltip = FALSE
)
}
\arguments{
\item{id}{\code{string} The layer's identifier must be unique for among all layers of the same
type for a map. Defaults to \code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}, but should be explicitly defined for
updatable layers in a shiny application.}

\item{name}{\code{string} The layer's name identifies the layer on tooltips and legends. It does
not need to be unique, but should be brief. Defaults to the deck.gl class name for the layer.}

\item{data}{data}

\item{visible}{boolean}

\item{pickable}{boolean}

\item{opacity}{number}

\item{position_format}{unknown}

\item{color_format}{unknown}

\item{auto_highlight}{boolean}

\item{highlight_color}{<\code{\link{accessor}}>}

\item{disk_resolution}{number}

\item{vertices}{unknown}

\item{radius}{number}

\item{angle}{number}

\item{offset}{array}

\item{coverage}{number}

\item{elevation_scale}{number}

\item{line_width_units}{unknown}

\item{line_width_scale}{number}

\item{line_width_min_pixels}{number}

\item{line_width_max_pixels}{number}

\item{extruded}{boolean}

\item{wireframe}{boolean}

\item{filled}{boolean}

\item{stroked}{boolean}

\item{get_position}{<\code{\link{accessor}}>}

\item{get_fill_color}{<\code{\link{accessor}}>}

\item{get_line_color}{<\code{\link{accessor}}>}

\item{get_line_width}{<\code{\link{accessor}}>}

\item{get_elevation}{<\code{\link{accessor}}>}

\item{material}{boolean}

\item{cell_size}{number}

\item{tooltip}{<\code{\link{tooltip}}> Defines the columns (and their order) that will be displayed in
the layer tooltip, if \code{pickable = TRUE}.}
}
\description{
Grid Cell Layer
}
\seealso{
\url{https://github.com/visgl/deck.gl/blob/v8.3.11/docs/api-reference/layers/grid-cell-layer.md}

Other layers: 
\code{\link{arc_layer}},
\code{\link{bitmap_layer}},
\code{\link{column_layer}},
\code{\link{contour_layer}},
\code{\link{cpu_grid_layer}},
\code{\link{geojson_layer}},
\code{\link{gpu_grid_layer}},
\code{\link{great_circle_layer}},
\code{\link{grid_layer}},
\code{\link{h3_cluster_layer}},
\code{\link{h3_hexagon_layer}},
\code{\link{heatmap_layer}},
\code{\link{hexagon_layer}},
\code{\link{icon_layer}},
\code{\link{line_layer}},
\code{\link{mvt_layer}},
\code{\link{path_layer}},
\code{\link{point_cloud_layer}},
\code{\link{polygon_layer}},
\code{\link{s2_layer}},
\code{\link{scatterplot_layer}},
\code{\link{screen_grid_layer}},
\code{\link{simple_mesh_layer}},
\code{\link{solid_polygon_layer}},
\code{\link{terrain_layer}},
\code{\link{text_layer}},
\code{\link{tile_3d_layer}},
\code{\link{tile_layer}},
\code{\link{trips_layer}}
}
\concept{layers}
