% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deckgl-layers-doc.R, R/deckgl-layers.R
\name{heatmap_layer}
\alias{heatmap_layer}
\alias{add_heatmap_layer}
\title{Heatmap Layer}
\usage{
add_heatmap_layer(
  rdeck,
  ...,
  id = NULL,
  name = "HeatmapLayer",
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  get_position = position,
  get_weight = 1,
  intensity = 1,
  radius_pixels = 50,
  color_range = c("#ffffb2", "#fed976", "#feb24c", "#fd8d3c", "#f03b20", "#bd0026"),
  threshold = 0.05,
  color_domain = NULL,
  tooltip = FALSE
)
}
\description{
Heatmap Layer
}
\seealso{
\url{https://github.com/visgl/deck.gl/blob/v8.3.11/docs/api-reference/aggregation-layers/heatmap-layer.md}

Other layers: 
\code{\link{arc_layer}},
\code{\link{bitmap_layer}},
\code{\link{column_layer}},
\code{\link{contour_layer}},
\code{\link{cpu_grid_layer}},
\code{\link{geojson_layer}},
\code{\link{gpu_grid_layer}},
\code{\link{great_circle_layer}},
\code{\link{grid_cell_layer}},
\code{\link{grid_layer}},
\code{\link{h3_cluster_layer}},
\code{\link{h3_hexagon_layer}},
\code{\link{hexagon_layer}},
\code{\link{icon_layer}},
\code{\link{line_layer}},
\code{\link{mvt_layer}},
\code{\link{path_layer}},
\code{\link{point_cloud_layer}},
\code{\link{polygon_layer}},
\code{\link{s2_layer}},
\code{\link{scatterplot_layer}},
\code{\link{screen_grid_layer}},
\code{\link{simple_mesh_layer}},
\code{\link{solid_polygon_layer}},
\code{\link{terrain_layer}},
\code{\link{text_layer}},
\code{\link{tile_3d_layer}},
\code{\link{tile_layer}},
\code{\link{trips_layer}}

Other aggregation-layers: 
\code{\link{contour_layer}},
\code{\link{cpu_grid_layer}},
\code{\link{gpu_grid_layer}},
\code{\link{grid_layer}},
\code{\link{hexagon_layer}},
\code{\link{screen_grid_layer}}
}
\concept{aggregation-layers}
\concept{layers}
