% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagon_layer.R
\name{hexagon_layer}
\alias{hexagon_layer}
\title{\href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/hexagon-layer.md}{HexagonLayer} deck.gl layer.}
\usage{
hexagon_layer(
  id = NULL,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  color_domain = NULL,
  color_range = list(c(255, 255, 178), c(254, 217, 118), c(254, 178, 76), c(253, 141,
    60), c(240, 59, 32), c(189, 0, 38)),
  get_color_value = NULL,
  get_color_weight = NULL,
  color_aggregation = "SUM",
  lower_percentile = 0,
  upper_percentile = 100,
  color_scale_type = "quantize",
  elevation_domain = NULL,
  elevation_range = c(0, 1000),
  get_elevation_value = NULL,
  get_elevation_weight = NULL,
  elevation_aggregation = "SUM",
  elevation_lower_percentile = 0,
  elevation_upper_percentile = 100,
  elevation_scale = 1,
  elevation_scale_type = "linear",
  radius = 1000,
  coverage = 1,
  extruded = FALSE,
  hexagon_aggregator = NULL,
  get_position = NULL,
  material = TRUE,
  ...
)
}
\arguments{
\item{id}{\code{\link{character}}
The id of the layer. Layer ids must be unique per layer \code{type} for deck.gl
to properly distinguish between them.}

\item{data}{\code{\link{data.frame}} | \code{\link[sf:sf]{sf::sf}}}

\item{visible}{\code{\link{logical}}}

\item{pickable}{\code{\link{logical}}}

\item{opacity}{\code{\link{numeric}}}

\item{position_format}{\code{XY} | \code{XYZ}}

\item{color_format}{\code{RGB} | \code{RGBA}}

\item{auto_highlight}{\code{\link{logical}}}

\item{highlight_color}{\code{\link{integer}}}

\item{color_domain}{\code{\link{numeric}}}

\item{color_range}{\code{\link{list}}}

\item{get_color_value}{accessor}

\item{get_color_weight}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{color_aggregation}{\code{SUM} | \code{MEAN} | \code{MIN} | \code{MAX}}

\item{lower_percentile}{\code{\link{numeric}}}

\item{upper_percentile}{\code{\link{numeric}}}

\item{color_scale_type}{\code{quantize} | \code{linear} | \code{quantile} | \code{ordinal}}

\item{elevation_domain}{\code{\link{numeric}}}

\item{elevation_range}{\code{\link{numeric}}}

\item{get_elevation_value}{accessor}

\item{get_elevation_weight}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{elevation_aggregation}{\code{SUM} | \code{MEAN} | \code{MIN} | \code{MAX}}

\item{elevation_lower_percentile}{\code{\link{numeric}}}

\item{elevation_upper_percentile}{\code{\link{numeric}}}

\item{elevation_scale}{\code{\link{numeric}}}

\item{elevation_scale_type}{\code{quantize} | \code{linear} | \code{quantile} | \code{ordinal}}

\item{radius}{\code{\link{numeric}}}

\item{coverage}{\code{\link{numeric}}}

\item{extruded}{\code{\link{logical}}}

\item{hexagon_aggregator}{\code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{get_position}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{material}{\code{\link{logical}}}

\item{...}{additional layer parameters to pass to deck.gl.
\code{snake_case} parameters will be converted to \code{snakeCase}.}
}
\value{
\code{HexagonLayer} & \code{\link{layer}}
A \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/hexagon-layer.md}{HexagonLayer} layer.
Add to an \link{rdeck} map via \code{\link{add_layer}} or \code{\link{rdeck}}.
}
\description{
\href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/hexagon-layer.md}{HexagonLayer} deck.gl layer.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/hexagon-layer.md}
}
