% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deckgl-layers.R, R/doc-deckgl-layers.R
\name{add_icon_layer}
\alias{add_icon_layer}
\alias{icon_layer}
\title{Icon Layer}
\usage{
add_icon_layer(
  rdeck,
  ...,
  id = NULL,
  name = "IconLayer",
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  icon_atlas = NULL,
  icon_mapping = list(),
  size_scale = 1,
  billboard = TRUE,
  size_units = "pixels",
  size_min_pixels = 0,
  size_max_pixels = 9007199254740991,
  alpha_cutoff = 0.05,
  get_position = position,
  get_icon = icon,
  get_color = "#000000ff",
  get_size = 1,
  get_angle = 0,
  get_pixel_offset = c(0, 0),
  blending_mode = "normal",
  tooltip = FALSE
)
}
\arguments{
\item{rdeck}{<\code{\link{rdeck}} | \code{\link{rdeck_proxy}}> An rdeck map instance.}

\item{...}{Additional parameters that will be forwarded to deck.gl javascript without
validation nor processing. All dots must be named and will be \code{camelCased} when serialised.
Layers will raise a warning when dots are used, warning class \code{rdeck_dots_nonempty}.}

\item{id}{<\code{string}> The layer's identifier must be unique for among all layers of the same
type for a map. Defaults to \code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}, but should be explicitly defined for
updatable layers in a shiny application.}

\item{name}{<\code{string}> Identifies the layer on tooltips and legends. It does
not need to be unique, but should be brief. Defaults to the deck.gl class name for the layer.}

\item{data}{<\code{data.frame} | \code{sf} | \code{string}> The layer's data. Data frames will contain all
columns that are referenced by the layer's accessors. Strings will be interpreted as a URL and
data will be retrieved dynamically in the browser.}

\item{visible}{<\code{logical}> Determines whether the layer is visible or not; also determines
whether any legend elements for the layer will be displayed.}

\item{pickable}{<\code{logical}> Determines if the layer responds to pointer / touch events.}

\item{opacity}{<\code{number}> Determines the layer's opacity.}

\item{position_format}{<\code{"XY"} | \code{"XYZ"}> Determines whether each coordinate has two (XY)
or three (XYZ) elements.}

\item{color_format}{<\code{"RGB"} | \code{"RGBA"}> Determines whether the alpha channel of the colours
will be ignored by accessors, making all colours opaque.}

\item{auto_highlight}{<\code{logical}> When \code{TRUE}, the current object \emph{hovered} by the cursor is
highlighted by \code{highlight_color}.}

\item{highlight_color}{<\code{\link{accessor}} | \code{\link{scale}} | \code{\link{color}}> When \code{auto_highlight} and
\code{pickable} are enabled, \code{highlight_color} determines the colour of the currently
highlighted object. If a single colour value is supplied, that colour will be used to
highlight all objects in the layer. Per-object highlighting is achieved with a colour scale,
or a \href{https://dplyr.tidyverse.org/articles/programming.html}{tidy-eval} column of colours.}

\item{icon_atlas}{object}

\item{icon_mapping}{object}

\item{size_scale}{<\code{number}> The size multiplier.}

\item{billboard}{<\code{logical}> If \code{TRUE}, the text label always faces the camera, otherwise it
faces up (z).}

\item{size_units}{<\code{"pixels"} | \code{"meters"}> The units of the size specified by
\code{get_size}.}

\item{size_min_pixels}{<\code{number}> The minimum size in pixels.}

\item{size_max_pixels}{<\code{number}> The maximum size in pixels.}

\item{alpha_cutoff}{<\code{number}> Discard pixels whose opacity is below this threshold.
A discarded pixel would create a "hole" in the icon that is not considered part of
the object.}

\item{get_position}{<\code{\link{accessor}}> The position geometry column, an
\code{sfc_POINT} column with CRS \href{http://epsg.io/4326}{EPSG:4326}. Supports
\href{https://dplyr.tidyverse.org/articles/programming.html}{tidy-eval}.}

\item{get_icon}{<\code{\link{accessor}}>}

\item{get_color}{<\code{\link{accessor}} | \code{\link{scale}} | \code{\link{color}}> The colour of each object.
Accepts a single colour value, a colour scale, or a
\href{https://dplyr.tidyverse.org/articles/programming.html}{tidy-eval} column of colours.}

\item{get_size}{<\code{\link{accessor}} | \code{\link{scale}} | \code{number}> The icon size of each text label,
in units specified by \code{size_units}.
Accepts a single numeric value, a numeric scale, or a
\href{https://dplyr.tidyverse.org/articles/programming.html}{tidy-eval} column of numbers.}

\item{get_angle}{<\code{\link{accessor}} | \code{number}> The rotating angle of each icon in degrees.
Accepts a single numeric value, or a
\href{https://dplyr.tidyverse.org/articles/programming.html}{tidy-eval} column of numbers.}

\item{get_pixel_offset}{<\code{\link{accessor}} | \code{number}>}

\item{blending_mode}{<\code{"normal"} | \code{"additive"} | \code{"subtractive"}> Sets the blending mode.
Blending modes:
\itemize{
\item \code{normal}: Normal blending doesn't alter colours of overlapping objects.
\item \code{additive}: Additive blending adds colours of overlapping objects. Useful for highlighting
dot density on dark maps.
\item \code{subtractive}: Subtractive blending darkens overlapping objects. Useful for highlighting
dot density on light maps.
}}

\item{tooltip}{<\code{\link{tooltip}}> Defines the columns (and their order) that will be displayed in
the layer tooltip, if \code{pickable == TRUE}. Supports <\link[dplyr:dplyr_tidy_select]{tidy-select}>
if a \code{data} is a \code{data.frame}. \code{sfc} columns are always removed.}
}
\description{
Icon Layer
}
\seealso{
\url{https://github.com/visgl/deck.gl/blob/v8.3.11/docs/api-reference/layers/icon-layer.md}

Other core-layers: 
\code{\link{add_arc_layer}()},
\code{\link{add_bitmap_layer}()},
\code{\link{add_column_layer}()},
\code{\link{add_geojson_layer}()},
\code{\link{add_grid_cell_layer}()},
\code{\link{add_line_layer}()},
\code{\link{add_path_layer}()},
\code{\link{add_point_cloud_layer}()},
\code{\link{add_polygon_layer}()},
\code{\link{add_scatterplot_layer}()},
\code{\link{add_solid_polygon_layer}()},
\code{\link{add_text_layer}()}

Other layers: 
\code{\link{add_arc_layer}()},
\code{\link{add_bitmap_layer}()},
\code{\link{add_column_layer}()},
\code{\link{add_contour_layer}()},
\code{\link{add_cpu_grid_layer}()},
\code{\link{add_geojson_layer}()},
\code{\link{add_gpu_grid_layer}()},
\code{\link{add_great_circle_layer}()},
\code{\link{add_grid_cell_layer}()},
\code{\link{add_grid_layer}()},
\code{\link{add_h3_cluster_layer}()},
\code{\link{add_h3_hexagon_layer}()},
\code{\link{add_heatmap_layer}()},
\code{\link{add_hexagon_layer}()},
\code{\link{add_line_layer}()},
\code{\link{add_mvt_layer}()},
\code{\link{add_path_layer}()},
\code{\link{add_point_cloud_layer}()},
\code{\link{add_polygon_layer}()},
\code{\link{add_s2_layer}()},
\code{\link{add_scatterplot_layer}()},
\code{\link{add_scenegraph_layer}()},
\code{\link{add_screen_grid_layer}()},
\code{\link{add_simple_mesh_layer}()},
\code{\link{add_solid_polygon_layer}()},
\code{\link{add_terrain_layer}()},
\code{\link{add_text_layer}()},
\code{\link{add_tile_3d_layer}()},
\code{\link{add_tile_layer}()},
\code{\link{add_trips_layer}()}
}
\concept{core-layers}
\concept{layers}
