% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer.R
\name{layer}
\alias{layer}
\title{Create a deck.gl layer}
\usage{
layer(
  type = NULL,
  id = NULL,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  ...
)
}
\arguments{
\item{type}{\code{{character}}
deck.gl layer type, one of \link{layer_types}}

\item{id}{\code{{character}}
The id of the layer. Layer ids must be unique per layer \code{type} for deck.gl
to properly distiguish between them.}

\item{data}{\code{{data.frame | sf}}
Deck.gl layer data, either a \link[base:data.frame]{base::data.frame} or \link[sf:sf]{sf::sf}.}

\item{visible}{\code{{logical}}
Whether the layer is visible.}

\item{pickable}{\code{{logical}}
Whether the layer responds to mouse pointer picking events.}

\item{opacity}{\code{{numeric}}
The opacity of the layer.}

\item{position_format}{\code{{"XY" | "XYZ"}}
Determines whether each coordinate has two (XY) or three (XYZ) numbers.}

\item{color_format}{\code{{"RGB" | "RGBA"}}
RGB will make the layer ignore the alpha channel of colours returned by
accessors. Opacity controlled by \code{opacity} is still applied.}

\item{auto_highlight}{\code{{logical}}
When \code{TRUE}, current object pointed by mouse pointer (when hovered over) is
highlighted with highlight_color. Requires \code{pickable} to be \code{TRUE}.}

\item{highlight_color}{\code{{RGB | RGBA}}
RGBA color to be used to render highlighted object. When 3 component (RGB)
array is specified, a default value of 255 is used for alpha.}

\item{...}{additional deck.gl layer parameters}
}
\value{
\\verb{\{layer\}\\}
}
\description{
Create a deck.gl layer
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/master/docs/api-reference/layer.md}
}
