% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_trans.R
\name{log_trans}
\alias{log_trans}
\title{Log transformation}
\usage{
log_trans(base = exp(1))
}
\arguments{
\item{base}{<\code{number}> The log base}
}
\description{
Applies a log transform on the input: \code{y = log(abs(x), b)}, where:
\itemize{
\item \code{x} is the input vector
\item \code{k} is the exponent
}
}
\details{
If \code{x} is negative, the result is multiplied by -1.

This transform assumes -- but doesn't enforce -- that the input range
doesn't cross zero. Transforming an input that crosses 0 will give predictable
output, but its inverse will not, due to unsigned 0 (i.e. -1 * log(1) == log(1)).

For inverse, the sign for input zeros will be assumed negative if there any negative
input, else positive.
}
\seealso{
Other transform: 
\code{\link{power_trans}()},
\code{\link{symlog_trans}()}
}
\concept{transform}
