% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mvt_url}
\alias{mvt_url}
\title{MVT URL}
\usage{
mvt_url(tileset_id)
}
\arguments{
\item{tileset_id}{A mapbox tileset identifier of the form:
\itemize{
\item \code{mapbox.mapbox-streets-v8}, or
\item \verb{mapbox://mapbox.mapbox-streets-v8}
}}
}
\description{
Make a mapbox vector tile url template
}
\seealso{
add_mvt_layer
}
