% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_layer.R
\name{path_layer}
\alias{path_layer}
\title{\href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/path-layer.md}{PathLayer} deck.gl layer.}
\usage{
path_layer(
  id = NULL,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  width_units = "meters",
  width_scale = 1,
  width_min_pixels = 0,
  width_max_pixels = 9007199254740991,
  rounded = FALSE,
  miter_limit = 4,
  billboard = FALSE,
  get_path = NULL,
  get_color = NULL,
  get_width = NULL,
  ...
)
}
\arguments{
\item{id}{\code{\link{character}}
The id of the layer. Layer ids must be unique per layer \code{type} for deck.gl
to properly distinguish between them.}

\item{data}{\code{\link{data.frame}} | \code{\link[sf:sf]{sf::sf}}}

\item{visible}{\code{\link{logical}}}

\item{pickable}{\code{\link{logical}}}

\item{opacity}{\code{\link{numeric}}}

\item{position_format}{\code{XY} | \code{XYZ}}

\item{color_format}{\code{RGB} | \code{RGBA}}

\item{auto_highlight}{\code{\link{logical}}}

\item{highlight_color}{\code{\link{integer}}}

\item{width_units}{\code{pixels} | \code{meters}}

\item{width_scale}{\code{\link{numeric}}}

\item{width_min_pixels}{\code{\link{numeric}}}

\item{width_max_pixels}{\code{\link{numeric}}}

\item{rounded}{\code{\link{logical}}}

\item{miter_limit}{\code{\link{numeric}}}

\item{billboard}{\code{\link{logical}}}

\item{get_path}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{get_color}{accessor}

\item{get_width}{accessor | \code{\link{numeric}}}

\item{...}{additional layer parameters to pass to deck.gl.
\code{snake_case} parameters will be converted to \code{snakeCase}.}
}
\value{
\code{PathLayer} & \code{\link{layer}}
A \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/path-layer.md}{PathLayer} layer.
Add to an \link{rdeck} map via \code{\link{add_layer}} or \code{\link{rdeck}}.
}
\description{
\href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/path-layer.md}{PathLayer} deck.gl layer.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/path-layer.md}
}
