% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deckgl-layers-doc.R, R/deckgl-layers.R
\name{point_cloud_layer}
\alias{point_cloud_layer}
\alias{add_point_cloud_layer}
\title{Point Cloud Layer}
\usage{
add_point_cloud_layer(
  rdeck,
  ...,
  id = NULL,
  name = "PointCloudLayer",
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  size_units = "pixels",
  point_size = 10,
  get_position = position,
  get_normal = c(0, 0, 1),
  get_color = "#000000ff",
  material = TRUE,
  tooltip = FALSE
)
}
\arguments{
\item{id}{string}

\item{name}{string}

\item{data}{data}

\item{visible}{boolean}

\item{pickable}{boolean}

\item{opacity}{number}

\item{position_format}{unknown}

\item{color_format}{unknown}

\item{auto_highlight}{boolean}

\item{highlight_color}{accessor}

\item{size_units}{unknown}

\item{point_size}{number}

\item{get_position}{accessor}

\item{get_normal}{accessor}

\item{get_color}{accessor}

\item{material}{boolean}

\item{tooltip}{accessor}
}
\description{
Point Cloud Layer
}
\seealso{
\url{https://github.com/visgl/deck.gl/blob/v8.3.11/docs/api-reference/layers/point-cloud-layer.md}

Other layers: 
\code{\link{arc_layer}},
\code{\link{bitmap_layer}},
\code{\link{column_layer}},
\code{\link{contour_layer}},
\code{\link{cpu_grid_layer}},
\code{\link{geojson_layer}},
\code{\link{gpu_grid_layer}},
\code{\link{great_circle_layer}},
\code{\link{grid_cell_layer}},
\code{\link{grid_layer}},
\code{\link{h3_cluster_layer}},
\code{\link{h3_hexagon_layer}},
\code{\link{heatmap_layer}},
\code{\link{hexagon_layer}},
\code{\link{icon_layer}},
\code{\link{line_layer}},
\code{\link{mvt_layer}},
\code{\link{path_layer}},
\code{\link{polygon_layer}},
\code{\link{s2_layer}},
\code{\link{scatterplot_layer}},
\code{\link{screen_grid_layer}},
\code{\link{simple_mesh_layer}},
\code{\link{solid_polygon_layer}},
\code{\link{terrain_layer}},
\code{\link{text_layer}},
\code{\link{tile_3d_layer}},
\code{\link{tile_layer}},
\code{\link{trips_layer}}
}
\concept{layers}
