% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdeck-package.R
\docType{package}
\name{rdeck-package}
\alias{rdeck-package}
\alias{_PACKAGE}
\title{rdeck: Deck.gl Widget}
\description{
Deck.gl widget for R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://anthonynorth.github.io/rdeck}
  \item \url{https://github.com/anthonynorth/rdeck}
  \item Report bugs at \url{https://github.com/anthonynorth/rdeck/issues}
}

}
\author{
\strong{Maintainer}: Anthony North \email{anthony.jl.north@gmail.com}

}
\keyword{internal}
