% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{rdeck}
\alias{rdeck}
\title{RDeck}
\usage{
rdeck(
  map_style = mapbox_dark(),
  theme = "kepler",
  initial_bounds = NULL,
  initial_view_state = view_state(center = c(0, 0), zoom = 1),
  controller = TRUE,
  picking_radius = 0,
  use_device_pixels = TRUE,
  blending_mode = "normal",
  layer_selector = FALSE,
  lazy_load = FALSE,
  width = NULL,
  height = NULL,
  id = NULL,
  ...
)
}
\arguments{
\item{map_style}{<\code{string}> The mapbox basemap style url.
See \url{https://docs.mapbox.com/api/maps/#mapbox-styles}}

\item{theme}{<\code{"kepler"} | \code{"light"}> The widget theme which alters the style of the
legend and tooltips.}

\item{initial_bounds}{<\code{\link[sf:st_bbox]{st_bbox}} | \code{\link[sf:sf]{sf}} | \code{\link[sf:sfc]{sfc}}>
The initial bounds of the map; overwrites \code{initial_view_state}.}

\item{initial_view_state}{<\code{\link{view_state}}> Defines the map position, zoom, bearing and pitch.}

\item{controller}{<\code{logical}> If \code{NULL} or \code{FALSE}, the map is not interactive.}

\item{picking_radius}{<\code{number}> Extra pixels around the pointer to include while picking;
useful when rendering objects that are difficult to hover, e.g. thin lines, small points, etc.}

\item{use_device_pixels}{<\code{logical} | \code{number}> Controls the resolution of drawing buffer used
for rendering.
\itemize{
\item \code{TRUE}: Resolution is defined by \code{window.devicePixelRatio}. On Retina/HD displays, this
resolution is usually twice as big as \verb{CSS pixels} resolution.
\item \code{FALSE}: \verb{CSS pixels} resolution is used for rendering.
\item \code{number}: Custom ratio (drawing buffer resolution to \verb{CSS pixel}) to determine drawing
buffer size. A value less than \code{1} uses resolution smaller than \verb{CSS pixels}, improving
rendering performance at the expense of image quality; a value greater than \code{1} improves
image quality at the expense of rendering performance.
}}

\item{blending_mode}{<\code{"normal"} | \code{"additive"} | \code{"subtractive"}> Sets the blending mode.
Blending modes:
\itemize{
\item \code{normal}: Normal blending doesn't alter colours of overlapping objects.
\item \code{additive}: Additive blending adds colours of overlapping objects. Useful for highlighting
dot density on dark maps.
\item \code{subtractive}: Subtractive blending darkens overlapping objects. Useful for highlighting
dot density on light maps.
}}

\item{layer_selector}{<\code{boolean}> If \code{TRUE}, the layer selector control will be enabled
and layers with \code{visibility_toggle = TRUE} may be toggled. If \code{FALSE}, the layer selector control
won't be rendered.}

\item{lazy_load}{<\code{boolean}> If \code{TRUE}, maps will be rendered when they are scrolled into
view and destroyed when they are scrolled out of view. If \code{FALSE}, maps will be rendered
when the page loads.}

\item{width}{<\code{number}> The width of the map canvas.}

\item{height}{<\code{number}> The height of the map canvas.}

\item{id}{<\code{string}> The map element id. Not used in shiny applications.}

\item{...}{Additional parameters that will be forwarded to deck.gl javascript without
validation nor processing. All dots must be named and will be \code{camelCased} when serialised.
A warning is raised when dots are used, warning class \code{rdeck_dots_nonempty}.}
}
\description{
Create a Deck.GL map. Rendering the mapbox basemap requires a mapbox account and
\href{mapbox_access_token}{mapbox access token}.
}
\seealso{
\url{https://github.com/visgl/deck.gl/blob/8.4-release/docs/api-reference/core/deck.md}
}
