% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{rdeck}
\alias{rdeck}
\title{Create rdeck widget}
\usage{
rdeck(
  mapbox_api_access_token = Sys.getenv("MAPBOX_ACCESS_TOKEN"),
  map_style = "mapbox://styles/mapbox/dark-v10",
  initial_bounds = NULL,
  initial_view_state = NULL,
  layers = NULL,
  controller = TRUE,
  picking_radius = 0,
  use_device_pixels = TRUE,
  width = NULL,
  height = NULL,
  elementId = NULL,
  ...
)
}
\arguments{
\item{mapbox_api_access_token}{\code{\link{character}}
The api access token to use mapbox tiles.}

\item{map_style}{\code{\link{character}}
A mapbox style url. \url{https://docs.mapbox.com/api/maps/#mapbox-styles}}

\item{initial_bounds}{\code{\link{bounds}}
The initial bounds of the map; overwrites \code{initial_view_state}.}

\item{initial_view_state}{\code{\link{view_state}}
The initial view state of the map. See \link{view_state} for details.}

\item{layers}{\code{\link{list}}
The list of deck.gl layers.}

\item{controller}{\code{\link{logical}}
If \code{NULL} or \code{FALSE}, the map is not interactive.}

\item{picking_radius}{\code{\link{numeric}}
Extra pixels around the pointer to include while picking.}

\item{use_device_pixels}{\code{\link{logical}} | \code{\link{numeric}}
Controls the resolution of drawing buffer used for rendering.}

\item{width}{\code{\link{numeric}}
Width of the map}

\item{height}{\code{\link{numeric}}
Height of the map}

\item{elementId}{\code{\link{character}}
element id for the map.}

\item{...}{additional parameters to pass to the \href{https://github.com/uber/deck.gl/blob/master/docs/api-reference/deck.md}{Deck}.}
}
\value{
\code{\link{rdeck}}
An rdeck widget for a deck.gl map
}
\description{
Create rdeck widget
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/master/docs/api-reference/deck.md}
}
