% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{rdeck_html}
\alias{rdeck_html}
\title{Manually create the container to have control of dependency order}
\usage{
rdeck_html(id, style, class, ...)
}
\arguments{
\item{id}{\code{\link{character}}}

\item{style}{\code{\link{character}}}

\item{class}{\code{\link{character}}}

\item{...}{required by htmlwidgets}
}
\value{
\code{\link{rdeck}}
}
\description{
Manually create the container to have control of dependency order
}
\keyword{internal}
