% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale}
\alias{scale}
\title{Scale}
\usage{
scale(type, domain = NULL, range = c(0, 1), value, legend = TRUE, ...)
}
\description{
Scale
}
\seealso{
Other scales: 
\code{\link{scale_linear}()},
\code{\link{scale_log}()},
\code{\link{scale_power}()},
\code{\link{scale_quantile}()},
\code{\link{scale_quantize}()}
}
\concept{scales}
\keyword{internal}
