% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_category}
\alias{scale_category}
\alias{scale_color_category}
\title{Scale category}
\usage{
scale_color_category(
  col,
  palette,
  unmapped_color = "#000000",
  levels = NULL,
  tick_format = NULL,
  legend = TRUE
)

scale_category(
  col,
  range = 0:1,
  unmapped_value = 0,
  levels = NULL,
  legend = TRUE
)
}
\arguments{
\item{col}{<\code{name} | \code{string}> The name of the column containing data to be scaled.
Can be either a named column (non-standard evaluation), or an expression evaluating a string.
Supports \href{https://dplyr.tidyverse.org/articles/programming.html}{tidy-eval}.}

\item{palette}{<\code{\link{color}}> The colour palette of the colour scale. Must be a vector of
rgb\link{a} hex strings of length >= 2.}

\item{unmapped_color}{<\code{\link{color}}> The colour representing unmapped levels.}

\item{levels}{<\code{factor} | \code{character} | \code{logical}> The category levels. If NULL, will be
populated from input data. The order of the levels is determined by \code{levels()} for factors
& \code{unique()} otherwise. Length of \code{levels} must equal \code{palette} or \code{range}, such that each
category level is assigned a \code{color} or \code{range} value.}

\item{tick_format}{<\code{function}> A function taking a vector of ticks returning formatted ticks.}

\item{legend}{<\code{logical}> Indicate whether the legend should be displayed for this scale.}

\item{range}{<\code{numeric}> The output range of the numeric scale. Must be be a numeric vector of
length >= 2.}

\item{unmapped_value}{<\code{number}> The value representing unmapped levels.}
}
\description{
Creates a categorical scale.
}
\seealso{
Other scales: 
\code{\link{scale_linear}()},
\code{\link{scale_log}()},
\code{\link{scale_power}()},
\code{\link{scale_quantile}()},
\code{\link{scale_quantize}()},
\code{\link{scale_threshold}()}
}
\concept{scales}
