% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_category}
\alias{scale_category}
\alias{scale_color_category}
\title{Scale category}
\usage{
scale_color_category(
  col,
  palette = scales::brewer_pal("div"),
  unmapped_color = "#000000",
  levels = NULL,
  unmapped_tick = NULL,
  tick_format = NULL,
  legend = TRUE
)

scale_category(
  col,
  range = 0:1,
  unmapped_value = 0,
  levels = NULL,
  legend = TRUE
)
}
\arguments{
\item{col}{<\code{name} | \code{string}> The name of the column containing data to be scaled.
Must be a valid input to \code{\link[rlang:defusing-advanced]{rlang::ensym()}}; either a named column (non-standard evaluation), a string.
Supports \href{https://dplyr.tidyverse.org/articles/programming.html}{tidy-eval}.}

\item{palette}{<\code{\link{color}}|\code{function}> The colour palette of the colour scale. Must be a:
\itemize{
\item vector of RGBA hex colours,
\item a palette generator function, taking a length parameter, or
\item a palette ramp created from \code{\link[scales:colour_ramp]{scales::colour_ramp()}}
}

A \code{\link[scales:colour_ramp]{scales::colour_ramp()}} interpolator is created from the input palette.}

\item{unmapped_color}{<\code{\link{color}}> The colour representing unmapped levels.}

\item{levels}{<\code{factor} | \code{character} | \code{logical}> The category levels. If \code{NULL}, will be
populated from input data. The order of the levels is determined by \code{levels()} for factors
& \code{unique()} otherwise.

If there are more levels than colours (or range values), the palette (or range) is interpolated.}

\item{unmapped_tick}{<\code{string}> The tick label of the unmapped category. If not \code{NULL} and
\code{legend == TRUE}, the unmapped category will appear at the bottom of the legend.}

\item{tick_format}{<\code{function}> A label function taking a vector of ticks returning formatted ticks.}

\item{legend}{<\code{boolean}> Indicate whether the legend should be displayed for this scale.}

\item{range}{<\code{numeric}> The output range of the numeric scale. Must be length >= 2.
A \code{\link[stats:approxfun]{stats::approxfun()}} interpolator is created from the input range.}

\item{unmapped_value}{<\code{number}> The value representing unmapped levels.}
}
\description{
Creates a categorical scale. Categorical scales map input values defined in the set of
\code{levels} to colours (or values). Input values not in the set of \code{levels} are assigned
\code{unmapped_color} (or \code{unmapped_value}).
}
\seealso{
Other scales: 
\code{\link{scale_linear}()},
\code{\link{scale_log}()},
\code{\link{scale_power}()},
\code{\link{scale_quantile}()},
\code{\link{scale_quantize}()},
\code{\link{scale_symlog}()},
\code{\link{scale_threshold}()}
}
\concept{scales}
