% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_linear}
\alias{scale_linear}
\alias{scale_color_linear}
\title{Scale linear}
\usage{
scale_color_linear(
  col,
  palette,
  na_color = "#000000",
  limits = NULL,
  breaks = NULL,
  n_ticks = NULL,
  tick_format = format_number,
  legend = TRUE
)

scale_linear(
  col,
  range = 0:1,
  na_value = 0,
  limits = NULL,
  breaks = NULL,
  legend = TRUE
)
}
\arguments{
\item{col}{<\code{name} | \code{string}> The name of the column containing data to be scaled.
Can be either a named column (non-standard evaluation), or an expression evaluating a string.
Supports \href{https://dplyr.tidyverse.org/articles/programming.html}{tidy-eval}.}

\item{palette}{<\code{\link{color}}> The colour palette of the colour scale. Must be a vector of
rgb\link{a} hex strings of length >= 2.}

\item{na_color}{<\code{\link{color}}> The colour value for \code{NA} input values.}

\item{limits}{<\code{c(min, max)}> The limits of the scale's input. If not null, must be a
numeric vector of length 2, representing \code{c(min, max)}. Values outside the range of
\code{limits} will be clamped to the limits of the scale.}

\item{breaks}{<\code{numeric}> The breaks of the scale, allowing to define a piecewise scale.
If not null, must be \code{length(palette) - 2} or \code{length(range) - 2}, such that each \code{break}
is mapped to a \code{palette} or \code{range} entry.

Each break will be present on the legend for colour scales.}

\item{n_ticks}{<\code{number}> The number of ticks to display on the legend. Includes the domain
of the scale. If \code{breaks} is not \code{NULL}, each break must map to a tick and gaps between
breaks must have the same number of ticks; i.e. \code{n_ticks} must be
\code{2 + length(breaks) + x * (length(breaks) + 1)}, where \code{x} is any positive integer.

Defaults to \code{2 + length(breaks)} when \code{breaks} is not \code{NULL}; \code{6} otherwise.}

\item{tick_format}{<\code{function}> A function taking a vector of ticks returning formatted ticks.}

\item{legend}{<\code{logical}> Indicate whether the legend should be displayed for this scale.}

\item{range}{<\code{numeric}> The output range of the numeric scale. Must be be a numeric vector of
length >= 2.}

\item{na_value}{<\code{number}> The output value for \code{NA} input values.}
}
\description{
Creates a continuous linear scale. The colour palette (or range) is linearly interpolated
between \code{limits} (or between \code{limits} and between \code{breaks} for piecewise scales).
}
\seealso{
Other scales: 
\code{\link{scale_category}()},
\code{\link{scale_log}()},
\code{\link{scale_power}()},
\code{\link{scale_quantile}()},
\code{\link{scale_quantize}()},
\code{\link{scale_threshold}()}
}
\concept{scales}
