% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_log}
\alias{scale_log}
\title{Log Scale}
\usage{
scale_log(domain = NULL, range = c(0, 1), base = 10, value, legend = TRUE)
}
\description{
Log Scale
}
\seealso{
Other scales: 
\code{\link{scale_linear}()},
\code{\link{scale_power}()},
\code{\link{scale_quantile}()},
\code{\link{scale_quantize}()},
\code{\link{scale}()}
}
\concept{scales}
