% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_quantile}
\alias{scale_quantile}
\title{Quantile Scale}
\usage{
scale_quantile(range = c(0, 1), value, legend = TRUE)
}
\description{
Quantile Scale
}
\seealso{
Other scales: 
\code{\link{scale_linear}()},
\code{\link{scale_log}()},
\code{\link{scale_power}()},
\code{\link{scale_quantize}()},
\code{\link{scale}()}
}
\concept{scales}
