% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_quantile}
\alias{scale_quantile}
\alias{scale_color_quantile}
\title{Scale quantile}
\usage{
scale_color_quantile(
  col,
  palette = scales::viridis_pal(),
  na_color = "#000000",
  probs = seq.int(0, 1, 0.25),
  data = NULL,
  tick_format = format_number,
  legend = TRUE
)

scale_quantile(
  col,
  range = 0:1,
  na_value = 0,
  probs = seq.int(0, 1, 0.25),
  data = NULL,
  legend = TRUE
)
}
\arguments{
\item{col}{<\code{name} | \code{string}> The name of the column containing data to be scaled.
Must be a valid input to \code{\link[rlang:defusing-advanced]{rlang::ensym()}}; either a named column (non-standard evaluation), a string.
Supports \href{https://dplyr.tidyverse.org/articles/programming.html}{tidy-eval}.}

\item{palette}{<\code{\link{color}}|\code{function}> The colour palette of the colour scale. Must be a:
\itemize{
\item vector of RGBA hex colours,
\item a palette generator function, taking a length parameter, or
\item a palette ramp created from \code{\link[scales:colour_ramp]{scales::colour_ramp()}}
}

A \code{\link[scales:colour_ramp]{scales::colour_ramp()}} interpolator is created from the input palette.}

\item{na_color}{<\code{\link{color}}> The colour value for \code{NA} input values.}

\item{probs}{<\code{numeric}> The quantile probabilities. Must be between 0 and 1.}

\item{data}{<\code{numeric}> The data used to compute the quantiles. If \code{NULL}, will be
taken from the layer data.}

\item{tick_format}{<\code{function}> A label function taking a vector of ticks returning formatted ticks.}

\item{legend}{<\code{boolean}> Indicate whether the legend should be displayed for this scale.}

\item{range}{<\code{numeric}> The output range of the numeric scale. Must be length >= 2.
A \code{\link[stats:approxfun]{stats::approxfun()}} interpolator is created from the input range.}

\item{na_value}{<\code{number}> The output value for \code{NA} input values.}
}
\description{
Creates a threshold scale, where threshold breaks are computed from the given quantile
\code{probs}.

Quantile scale legend ticks will be quantile values at each quantile break (including
limits), not the quantile probabilities at each break. This can be overridden in
\code{tick_format}.

Example:
\code{tick_format = function(x) as.character(probs)}.
}
\note{
As the quantiles are computed from input data, quantile scales are incompatible with
layers that load data from a url (e.g \code{mvt_layer}). If quantiles for remote data are
known, a quantile scale can be constructed manually with \code{\link[=scale_threshold]{scale_threshold()}}.
}
\seealso{
Other scales: 
\code{\link{scale_category}()},
\code{\link{scale_linear}()},
\code{\link{scale_log}()},
\code{\link{scale_power}()},
\code{\link{scale_quantize}()},
\code{\link{scale_symlog}()},
\code{\link{scale_threshold}()}
}
\concept{scales}
