% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_quantile}
\alias{scale_quantile}
\alias{scale_color_quantile}
\title{Scale quantile}
\usage{
scale_color_quantile(
  col,
  palette,
  na_color = "#000000",
  tick_format = format_number,
  legend = TRUE
)

scale_quantile(col, range = 1:5, na_value = 0, legend = FALSE)
}
\arguments{
\item{col}{<\code{name} | \code{string}> The name of the column containing data to be scaled.
Can be either a named column (non-standard evaluation), or an expression evaluating a string.
Supports \href{https://dplyr.tidyverse.org/articles/programming.html}{tidy-eval}.}

\item{palette}{<\code{color}> The colour palette of the colour scale. Must be a vector of
rgb\link{a} hex strings of length >= 2.}

\item{na_color}{<\code{color}> The colour value for \code{NA} input values.}

\item{tick_format}{<\code{function}> A function taking a vector of ticks returning formatted ticks.}

\item{legend}{<\code{logical}> Indicate whether the legend should be displayed for this scale.}

\item{range}{<\code{numeric}> The output range of the numeric scale. Must be be a numeric vector of
length >= 2.}

\item{na_value}{<\code{number}> The output value for \code{NA} input values.}
}
\description{
Creates a quantile scale. Number of quantiles is defined by the length of
\code{palette} or \code{range}.
}
\note{
As the quantiles are computed from input data, quantile scales are incompatible with
layers that load data from a url (e.g \code{mvt_layer}). If quantiles for remote data are
known, a quantile scale can be constructed manually with \code{\link[=scale_threshold]{scale_threshold()}}.
}
\seealso{
Other scales: 
\code{\link{scale_category}()},
\code{\link{scale_linear}()},
\code{\link{scale_log}()},
\code{\link{scale_power}()},
\code{\link{scale_quantize}()},
\code{\link{scale_threshold}()}
}
\concept{scales}
