% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_quantize}
\alias{scale_quantize}
\title{Quantize Scale}
\usage{
scale_quantize(domain = NULL, range = c(0, 1), value, legend = TRUE)
}
\description{
Quantize Scale
}
\seealso{
Other scales: 
\code{\link{scale_linear}()},
\code{\link{scale_log}()},
\code{\link{scale_power}()},
\code{\link{scale_quantile}()},
\code{\link{scale}()}
}
\concept{scales}
