% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_quantize}
\alias{scale_quantize}
\alias{scale_color_quantize}
\title{Scale quantize}
\usage{
scale_color_quantize(
  col,
  palette = scales::viridis_pal(),
  na_color = "#000000",
  limits = NULL,
  n_breaks = 6,
  tick_format = format_number,
  col_label = "{.col}",
  legend = TRUE
)

scale_quantize(
  col,
  range = 0:1,
  na_value = 0,
  limits = NULL,
  n_breaks = 6,
  col_label = "{.col}",
  legend = TRUE
)
}
\arguments{
\item{col}{<\code{name} | \code{string}> The name of the column containing data to be scaled.
Must be a valid input to \code{\link[rlang:defusing-advanced]{rlang::ensym()}}; either a named column (non-standard evaluation), a string.
Supports \href{https://dplyr.tidyverse.org/articles/programming.html}{tidy-eval}.}

\item{palette}{<\code{\link{color}}|\code{function}> The colour palette of the colour scale. Must be a:
\itemize{
\item vector of RGBA hex colours,
\item a palette generator function, taking a length parameter, or
\item a palette ramp created from \code{\link[scales:colour_ramp]{scales::colour_ramp()}}
}

A \code{\link[scales:colour_ramp]{scales::colour_ramp()}} interpolator is created from the input palette.}

\item{na_color}{<\code{\link{color}}> The colour value for \code{NA} input values.}

\item{limits}{<\code{c(min, max)}> The limits of the scale's input. If \code{NULL}, limits are computed
from layer data. Values outside the range of limits are clamped.}

\item{n_breaks}{<\code{integer}> The number of linearly spaced breaks in the scale.
Each break will be present on the legend for colour scales.}

\item{tick_format}{<\code{function}> A label function taking a vector of ticks returning formatted ticks.}

\item{col_label}{<\code{string} | \code{function}> A template string or a label function for customising the
column name in the legend.
\itemize{
\item if \code{col_label} is a string, \code{{.col}} may be used to represent the \code{col} name
\item if \code{col_label} is a function, the function must take a single argument: the \code{col} name
}}

\item{legend}{<\code{boolean}> Indicate whether the legend should be displayed for this scale.}

\item{range}{<\code{numeric}> The output range of the numeric scale. Must be length >= 2.
A \code{\link[stats:approxfun]{stats::approxfun()}} interpolator is created from the input range.}

\item{na_value}{<\code{number}> The output value for \code{NA} input values.}
}
\description{
Creates a discrete quantize scale, with \code{n_breaks} linearly spaced threshold breaks between
\code{limits}. This scale can be thought of as a restricted special case of \code{\link[=scale_threshold]{scale_threshold()}},
or a discrete \code{\link[=scale_linear]{scale_linear()}}.
}
\seealso{
Other scales: 
\code{\link{scale_category}()},
\code{\link{scale_identity}()},
\code{\link{scale_linear}()},
\code{\link{scale_log}()},
\code{\link{scale_power}()},
\code{\link{scale_quantile}()},
\code{\link{scale_symlog}()},
\code{\link{scale_threshold}()}
}
\concept{scales}
