% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deckgl-layers-doc.R, R/deckgl-layers.R
\name{scenegraph_layer}
\alias{scenegraph_layer}
\alias{add_scenegraph_layer}
\title{Scenegraph Layer}
\usage{
add_scenegraph_layer(
  rdeck,
  ...,
  id = NULL,
  name = "ScenegraphLayer",
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  scenegraph = NULL,
  get_scene = NULL,
  get_animator = NULL,
  size_scale = 1,
  size_min_pixels = 0,
  size_max_pixels = 9007199254740991,
  get_position = position,
  get_color = "#ffffffff",
  get_orientation = c(0, 0, 0),
  get_scale = c(1, 1, 1),
  get_translation = c(0, 0, 0),
  get_transform_matrix = NULL,
  tooltip = FALSE
)
}
\description{
Scenegraph Layer
}
\seealso{
\url{https://github.com/visgl/deck.gl/blob/v8.3.11/docs/api-reference/mesh-layers/scenegraph-layer.md}

Other mesh-layers: 
\code{\link{simple_mesh_layer}}
}
\concept{mesh-layers}
