% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deckgl-layers-doc.R, R/deckgl-layers.R
\name{scenegraph_layer}
\alias{scenegraph_layer}
\alias{add_scenegraph_layer}
\title{Scenegraph Layer}
\usage{
add_scenegraph_layer(
  rdeck,
  ...,
  id = NULL,
  name = "ScenegraphLayer",
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  scenegraph = NULL,
  get_scene = NULL,
  get_animator = NULL,
  size_scale = 1,
  size_min_pixels = 0,
  size_max_pixels = 9007199254740991,
  get_position = position,
  get_color = "#ffffffff",
  get_orientation = c(0, 0, 0),
  get_scale = c(1, 1, 1),
  get_translation = c(0, 0, 0),
  get_transform_matrix = NULL,
  tooltip = FALSE
)
}
\arguments{
\item{id}{\code{string} The layer's identifier must be unique for among all layers of the same
type for a map. Defaults to \code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}, but should be explicitly defined for
updatable layers in a shiny application.}

\item{name}{\code{string} The layer's name identifies the layer on tooltips and legends. It does
not need to be unique, but should be brief. Defaults to the deck.gl class name for the layer.}

\item{data}{data}

\item{visible}{boolean}

\item{pickable}{boolean}

\item{opacity}{number}

\item{position_format}{unknown}

\item{color_format}{unknown}

\item{auto_highlight}{boolean}

\item{highlight_color}{<\code{\link{accessor}}>}

\item{scenegraph}{object}

\item{get_scene}{function}

\item{get_animator}{function}

\item{size_scale}{number}

\item{size_min_pixels}{number}

\item{size_max_pixels}{number}

\item{get_position}{<\code{\link{accessor}}>}

\item{get_color}{<\code{\link{accessor}}>}

\item{get_orientation}{<\code{\link{accessor}}>}

\item{get_scale}{<\code{\link{accessor}}>}

\item{get_translation}{<\code{\link{accessor}}>}

\item{get_transform_matrix}{<\code{\link{accessor}}>}

\item{tooltip}{<\code{\link{tooltip}}> Defines the columns (and their order) that will be displayed in
the layer tooltip, if \code{pickable = TRUE}.}
}
\description{
Scenegraph Layer
}
\seealso{
\url{https://github.com/visgl/deck.gl/blob/v8.3.11/docs/api-reference/mesh-layers/scenegraph-layer.md}

Other mesh-layers: 
\code{\link{simple_mesh_layer}}
}
\concept{mesh-layers}
