% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{split_geometry}
\alias{split_geometry}
\title{split_geometry}
\usage{
split_geometry(obj)
}
\arguments{
\item{obj}{\code{{sf}} sf object}
}
\value{
\code{{sf}}
}
\description{
Split multi-geometries for deck.gl compatibility
}
