% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deckgl-layers-doc.R, R/deckgl-layers.R
\name{terrain_layer}
\alias{terrain_layer}
\alias{add_terrain_layer}
\title{Terrain Layer}
\usage{
add_terrain_layer(
  rdeck,
  ...,
  id = NULL,
  name = "TerrainLayer",
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  get_tile_data = NULL,
  extent = NULL,
  tile_size = 512,
  max_zoom = NULL,
  min_zoom = 0,
  max_cache_size = NULL,
  max_cache_byte_size = NULL,
  refinement_strategy = "best-available",
  z_range = NULL,
  max_requests = 6,
  elevation_data = "",
  texture = "",
  mesh_max_error = 4,
  bounds = NULL,
  color = c(255, 255, 255),
  elevation_decoder = list(rScaler = 1, gScaler = 0, bScaler = 0, offset = 0),
  worker_url = NULL,
  wireframe = FALSE,
  material = TRUE,
  tooltip = FALSE
)
}
\arguments{
\item{id}{\code{string} The layer's identifier must be unique for among all layers of the same
type for a map. Defaults to \code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}, but should be explicitly defined for
updatable layers in a shiny application.}

\item{name}{\code{string} The layer's name identifies the layer on tooltips and legends. It does
not need to be unique, but should be brief. Defaults to the deck.gl class name for the layer.}

\item{data}{data}

\item{visible}{boolean}

\item{pickable}{boolean}

\item{opacity}{number}

\item{position_format}{unknown}

\item{color_format}{unknown}

\item{auto_highlight}{boolean}

\item{highlight_color}{<\code{\link{accessor}}>}

\item{get_tile_data}{function}

\item{extent}{array}

\item{tile_size}{number}

\item{max_zoom}{unknown}

\item{min_zoom}{number}

\item{max_cache_size}{unknown}

\item{max_cache_byte_size}{unknown}

\item{refinement_strategy}{unknown}

\item{z_range}{unknown}

\item{max_requests}{number}

\item{elevation_data}{url}

\item{texture}{url}

\item{mesh_max_error}{number}

\item{bounds}{array}

\item{color}{color}

\item{elevation_decoder}{object}

\item{worker_url}{string}

\item{wireframe}{boolean}

\item{material}{boolean}

\item{tooltip}{<\code{\link{tooltip}}> Defines the columns (and their order) that will be displayed in
the layer tooltip, if \code{pickable = TRUE}.}
}
\description{
Terrain Layer
}
\seealso{
\url{https://github.com/visgl/deck.gl/blob/v8.3.11/docs/api-reference/geo-layers/terrain-layer.md}

Other layers: 
\code{\link{arc_layer}},
\code{\link{bitmap_layer}},
\code{\link{column_layer}},
\code{\link{contour_layer}},
\code{\link{cpu_grid_layer}},
\code{\link{geojson_layer}},
\code{\link{gpu_grid_layer}},
\code{\link{great_circle_layer}},
\code{\link{grid_cell_layer}},
\code{\link{grid_layer}},
\code{\link{h3_cluster_layer}},
\code{\link{h3_hexagon_layer}},
\code{\link{heatmap_layer}},
\code{\link{hexagon_layer}},
\code{\link{icon_layer}},
\code{\link{line_layer}},
\code{\link{mvt_layer}},
\code{\link{path_layer}},
\code{\link{point_cloud_layer}},
\code{\link{polygon_layer}},
\code{\link{s2_layer}},
\code{\link{scatterplot_layer}},
\code{\link{screen_grid_layer}},
\code{\link{simple_mesh_layer}},
\code{\link{solid_polygon_layer}},
\code{\link{text_layer}},
\code{\link{tile_3d_layer}},
\code{\link{tile_layer}},
\code{\link{trips_layer}}

Other geo-layers: 
\code{\link{great_circle_layer}},
\code{\link{h3_cluster_layer}},
\code{\link{h3_hexagon_layer}},
\code{\link{mvt_layer}},
\code{\link{s2_layer}},
\code{\link{tile_3d_layer}},
\code{\link{tile_layer}},
\code{\link{trips_layer}}
}
\concept{geo-layers}
\concept{layers}
