% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_layer.R
\name{text_layer}
\alias{text_layer}
\alias{add_text_layer}
\title{TextLayer}
\usage{
text_layer(
  id = "TextLayer",
  data = data.frame(),
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  billboard = TRUE,
  size_scale = 1,
  size_units = "pixels",
  size_min_pixels = 0,
  size_max_pixels = 9007199254740991,
  background_color = NULL,
  character_set = c(" ", "!", "\\"", "#", "$", "\%", "&", "'", "(", ")", "*", "+", ",",
    "-", ".", "/", "0", "1", "2", "3", "4", "5", "6", "7", "8", "9", ":", ";", "<", "=",
    ">", "?", "@", "A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L", "M", "N",
    "O", "P", "Q", "R", "S", "T", "U", "V", "W", "X", "Y", "Z", "[", "\\\\", "]", "^",
    "_", "`", "a", "b", "c", "d", "e", "f", "g", "h", "i", "j", "k", "l", "m", "n", "o",
    "p", "q", "r", "s", "t", "u", "v", "w", "x", "y", "z", "{", "|", "}", "~", "\\177"),
  font_family = "Monaco, monospace",
  font_weight = "normal",
  line_height = 1,
  font_settings = NULL,
  word_break = "word-break",
  max_width = -1,
  get_text = text,
  get_position = position,
  get_color = "#000000ff",
  get_size = 32,
  get_angle = 0,
  get_text_anchor = "middle",
  get_alignment_baseline = "center",
  get_pixel_offset = c(0, 0),
  ...
)

add_text_layer(
  rdeck,
  id = "TextLayer",
  data = data.frame(),
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  billboard = TRUE,
  size_scale = 1,
  size_units = "pixels",
  size_min_pixels = 0,
  size_max_pixels = 9007199254740991,
  background_color = NULL,
  character_set = c(" ", "!", "\\"", "#", "$", "\%", "&", "'", "(", ")", "*", "+", ",",
    "-", ".", "/", "0", "1", "2", "3", "4", "5", "6", "7", "8", "9", ":", ";", "<", "=",
    ">", "?", "@", "A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L", "M", "N",
    "O", "P", "Q", "R", "S", "T", "U", "V", "W", "X", "Y", "Z", "[", "\\\\", "]", "^",
    "_", "`", "a", "b", "c", "d", "e", "f", "g", "h", "i", "j", "k", "l", "m", "n", "o",
    "p", "q", "r", "s", "t", "u", "v", "w", "x", "y", "z", "{", "|", "}", "~", "\\177"),
  font_family = "Monaco, monospace",
  font_weight = "normal",
  line_height = 1,
  font_settings = NULL,
  word_break = "word-break",
  max_width = -1,
  get_text = text,
  get_position = position,
  get_color = "#000000ff",
  get_size = 32,
  get_angle = 0,
  get_text_anchor = "middle",
  get_alignment_baseline = "center",
  get_pixel_offset = c(0, 0),
  ...
)
}
\arguments{
\item{id}{\code{character}
The id of the layer. Layer ids must be unique per layer \code{type} for deck.gl
to properly distinguish between them.}

\item{data}{\code{data.frame} | \code{sf::sf}
Deck.gl layer data, either a \link{data.frame} or \link[sf:sf]{sf::sf}.}

\item{visible}{\code{logical}
Whether the layer is visible.}

\item{pickable}{\code{logical}
Whether the layer responds to mouse pointer picking events.}

\item{opacity}{\code{numeric}
The opacity of the layer.}

\item{position_format}{\code{"XY"} | \code{"XYZ"}
Determines whether each coordinate has two (XY) or three (XYZ) numbers.}

\item{color_format}{\code{"RGB"} | \code{"RGBA"}
RGB will make the layer ignore the alpha channel of colours returned by
accessors. Opacity controlled by \code{opacity} is still applied.}

\item{auto_highlight}{\code{logical}
When \code{TRUE}, current object pointed by mouse pointer (when hovered over) is
highlighted with highlight_color. Requires \code{pickable} to be \code{TRUE}.}

\item{highlight_color}{\code{integer} vector of (R,G,B) or (R,G,B,A)
RGBA color to be used to render highlighted object. When 3 component (RGB)
array is specified, a default value of 255 is used for alpha.}

\item{billboard}{\code{description-placeholder}}

\item{size_scale}{\code{description-placeholder}}

\item{size_units}{\code{description-placeholder}}

\item{size_min_pixels}{\code{description-placeholder}}

\item{size_max_pixels}{\code{description-placeholder}}

\item{background_color}{\code{description-placeholder}}

\item{character_set}{\code{description-placeholder}}

\item{font_family}{\code{description-placeholder}}

\item{font_weight}{\code{description-placeholder}}

\item{line_height}{\code{description-placeholder}}

\item{font_settings}{\code{description-placeholder}}

\item{word_break}{\code{description-placeholder}}

\item{max_width}{\code{description-placeholder}}

\item{get_text}{\code{description-placeholder}}

\item{get_position}{\code{description-placeholder}}

\item{get_color}{\code{description-placeholder}}

\item{get_size}{\code{description-placeholder}}

\item{get_angle}{\code{description-placeholder}}

\item{get_text_anchor}{\code{description-placeholder}}

\item{get_alignment_baseline}{\code{description-placeholder}}

\item{get_pixel_offset}{\code{description-placeholder}}

\item{...}{additional deck.gl layer parameters}

\item{rdeck}{\code{rdeck}
An rdeck map}
}
\description{
TextLayer
}
\section{Functions}{
\itemize{
\item \code{add_text_layer}: Add TextLayer to an rdeck map
}}

\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.1.0/docs/layers/text-layer.md}

Other layers: 
\code{\link{arc_layer}()},
\code{\link{bitmap_layer}()},
\code{\link{column_layer}()},
\code{\link{contour_layer}()},
\code{\link{cpu_grid_layer}()},
\code{\link{geojson_layer}()},
\code{\link{gpu_grid_layer}()},
\code{\link{great_circle_layer}()},
\code{\link{grid_cell_layer}()},
\code{\link{grid_layer}()},
\code{\link{h3_cluster_layer}()},
\code{\link{h3_hexagon_layer}()},
\code{\link{heatmap_layer}()},
\code{\link{hexagon_layer}()},
\code{\link{icon_layer}()},
\code{\link{line_layer}()},
\code{\link{mvt_layer}()},
\code{\link{path_layer}()},
\code{\link{point_cloud_layer}()},
\code{\link{polygon_layer}()},
\code{\link{s2_layer}()},
\code{\link{scatterplot_layer}()},
\code{\link{scenegraph_layer}()},
\code{\link{screen_grid_layer}()},
\code{\link{simple_mesh_layer}()},
\code{\link{solid_polygon_layer}()},
\code{\link{terrain_layer}()},
\code{\link{tile3d_layer}()},
\code{\link{tile_layer}()},
\code{\link{trips_layer}()}
}
\concept{layers}
