% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_layer.R
\name{text_layer}
\alias{text_layer}
\title{\href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/text-layer.md}{TextLayer} deck.gl layer.}
\usage{
text_layer(
  id = NULL,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  billboard = TRUE,
  size_scale = 1,
  size_units = "pixels",
  size_min_pixels = 0,
  size_max_pixels = 9007199254740991,
  background_color = NULL,
  character_set = NULL,
  font_family = "Monaco, monospace",
  font_weight = "normal",
  line_height = 1,
  font_settings = NULL,
  word_break = "word-break",
  max_width = -1,
  get_text = NULL,
  get_position = NULL,
  get_color = NULL,
  get_size = NULL,
  get_angle = NULL,
  get_text_anchor = NULL,
  get_alignment_baseline = NULL,
  get_pixel_offset = NULL,
  ...
)
}
\arguments{
\item{id}{\code{\link{character}}
The id of the layer. Layer ids must be unique per layer \code{type} for deck.gl
to properly distinguish between them.}

\item{data}{\code{\link{data.frame}} | \code{\link[sf:sf]{sf::sf}}}

\item{visible}{\code{\link{logical}}}

\item{pickable}{\code{\link{logical}}}

\item{opacity}{\code{\link{numeric}}}

\item{position_format}{\code{XY} | \code{XYZ}}

\item{color_format}{\code{RGB} | \code{RGBA}}

\item{auto_highlight}{\code{\link{logical}}}

\item{highlight_color}{\code{\link{integer}}}

\item{billboard}{\code{\link{logical}}}

\item{size_scale}{\code{\link{numeric}}}

\item{size_units}{\code{pixels} | \code{meters}}

\item{size_min_pixels}{\code{\link{numeric}}}

\item{size_max_pixels}{\code{\link{numeric}}}

\item{background_color}{\code{\link{integer}}}

\item{character_set}{\verb{[character]}}

\item{font_family}{\code{\link{character}}}

\item{font_weight}{\code{\link{character}}}

\item{line_height}{\code{\link{numeric}}}

\item{font_settings}{\code{\link{list}}}

\item{word_break}{\code{break-all} | \code{break-word}}

\item{max_width}{\code{\link{numeric}}}

\item{get_text}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{get_position}{accessor | \code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{get_color}{accessor}

\item{get_size}{accessor | \code{\link{numeric}}}

\item{get_angle}{accessor | \code{\link{numeric}}}

\item{get_text_anchor}{accessor | \verb{[character]}}

\item{get_alignment_baseline}{accessor | \verb{[character]}}

\item{get_pixel_offset}{accessor}

\item{...}{additional layer parameters to pass to deck.gl.
\code{snake_case} parameters will be converted to \code{snakeCase}.}
}
\value{
\code{TextLayer} & \code{\link{layer}}
A \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/text-layer.md}{TextLayer} layer.
Add to an \link{rdeck} map via \code{\link{add_layer}} or \code{\link{rdeck}}.
}
\description{
\href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/text-layer.md}{TextLayer} deck.gl layer.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/text-layer.md}
}
