% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deckgl-layers-doc.R, R/deckgl-layers.R
\name{text_layer}
\alias{text_layer}
\alias{add_text_layer}
\title{Text Layer}
\usage{
add_text_layer(
  rdeck,
  ...,
  id = NULL,
  name = "TextLayer",
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  billboard = TRUE,
  size_scale = 1,
  size_units = "pixels",
  size_min_pixels = 0,
  size_max_pixels = 9007199254740991,
  background_color = NULL,
  font_family = "Roboto, Helvetica, Arial, san-serif",
  font_weight = "normal",
  line_height = 1,
  font_settings = list(),
  word_break = "word-break",
  max_width = -1,
  get_text = text,
  get_position = position,
  get_color = "#000000ff",
  get_size = 32,
  get_angle = 0,
  get_text_anchor = "middle",
  get_alignment_baseline = "center",
  get_pixel_offset = c(0, 0),
  tooltip = FALSE
)
}
\arguments{
\item{id}{<\code{string}> The layer's identifier must be unique for among all layers of the same
type for a map. Defaults to \code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}, but should be explicitly defined for
updatable layers in a shiny application.}

\item{name}{<\code{string}> Identifies the layer on tooltips and legends. It does
not need to be unique, but should be brief. Defaults to the deck.gl class name for the layer.}

\item{data}{<\code{data.frame} | \code{sf} | \code{string}> The layer's data. Data frames will contain all
columns that are referenced by the layer's accessors. Strings will be interpreted as a URL and
data will be retrieved dynamically in the browser.}

\item{visible}{<\code{logical}> Determines whether the layer is visible or not; also determines
whether any legend elements for the layer will be displayed.}

\item{pickable}{<\code{logical}> Determines if the layer responds to pointer / touch events.}

\item{opacity}{<\code{numeric}> Determines the layer's opacity.}

\item{position_format}{<\code{"XY"} | \code{"XYZ"}> Determines whether each coordinate has two (XY)
or three (XYZ) elements.}

\item{color_format}{<\code{"RGB"} | \code{"RGBA"}> Determines whether the alpha channel of the colours
will be ignored by accessors, making all colours opaque.}

\item{auto_highlight}{<\code{logical}> When \code{TRUE}, the current object \emph{hovered} by the cursor is
highlighted by \code{highlight_color}.}

\item{highlight_color}{<\code{\link{accessor}} | \code{\link{scale}} | \code{\link{color}}> When \code{auto_highlight} and
\code{pickable} are enabled, \code{highlight_color} determines the colour of the currently highlighted
object. If a single colour value is supplied, that colour will be used to highlight all objects
in the layer. Per-object highlighting is achieved with an \code{accessor} resolving a column in \code{data},
or a \code{scale} to transform a column into a colour.}

\item{billboard}{boolean}

\item{size_scale}{number}

\item{size_units}{unknown}

\item{size_min_pixels}{number}

\item{size_max_pixels}{number}

\item{background_color}{color}

\item{font_family}{unknown}

\item{font_weight}{unknown}

\item{line_height}{number}

\item{font_settings}{object}

\item{word_break}{unknown}

\item{max_width}{number}

\item{get_text}{<\code{\link{accessor}}>}

\item{get_position}{<\code{\link{accessor}}>}

\item{get_color}{<\code{\link{accessor}}>}

\item{get_size}{<\code{\link{accessor}}>}

\item{get_angle}{<\code{\link{accessor}}>}

\item{get_text_anchor}{<\code{\link{accessor}}>}

\item{get_alignment_baseline}{<\code{\link{accessor}}>}

\item{get_pixel_offset}{<\code{\link{accessor}}>}

\item{tooltip}{<\code{\link{tooltip}}> Defines the columns (and their order) that will be displayed in
the layer tooltip, if \code{pickable = TRUE}.}
}
\description{
Text Layer
}
\seealso{
\url{https://github.com/visgl/deck.gl/blob/v8.3.11/docs/api-reference/layers/text-layer.md}

Other layers: 
\code{\link{arc_layer}},
\code{\link{bitmap_layer}},
\code{\link{column_layer}},
\code{\link{contour_layer}},
\code{\link{cpu_grid_layer}},
\code{\link{geojson_layer}},
\code{\link{gpu_grid_layer}},
\code{\link{great_circle_layer}},
\code{\link{grid_cell_layer}},
\code{\link{grid_layer}},
\code{\link{h3_cluster_layer}},
\code{\link{h3_hexagon_layer}},
\code{\link{heatmap_layer}},
\code{\link{hexagon_layer}},
\code{\link{icon_layer}},
\code{\link{line_layer}},
\code{\link{mvt_layer}},
\code{\link{path_layer}},
\code{\link{point_cloud_layer}},
\code{\link{polygon_layer}},
\code{\link{s2_layer}},
\code{\link{scatterplot_layer}},
\code{\link{screen_grid_layer}},
\code{\link{simple_mesh_layer}},
\code{\link{solid_polygon_layer}},
\code{\link{terrain_layer}},
\code{\link{tile_3d_layer}},
\code{\link{tile_layer}},
\code{\link{trips_layer}}
}
\concept{layers}
