% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deckgl-layers-doc.R, R/deckgl-layers.R
\name{text_layer}
\alias{text_layer}
\alias{add_text_layer}
\title{Text Layer}
\usage{
add_text_layer(
  rdeck,
  ...,
  id = NULL,
  name = "TextLayer",
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  billboard = TRUE,
  size_scale = 1,
  size_units = "pixels",
  size_min_pixels = 0,
  size_max_pixels = 9007199254740991,
  background_color = NULL,
  font_family = "Roboto, Helvetica, Arial, san-serif",
  font_weight = "normal",
  line_height = 1,
  font_settings = list(),
  word_break = "word-break",
  max_width = -1,
  get_text = text,
  get_position = position,
  get_color = "#000000ff",
  get_size = 32,
  get_angle = 0,
  get_text_anchor = "middle",
  get_alignment_baseline = "center",
  get_pixel_offset = c(0, 0),
  tooltip = FALSE
)
}
\arguments{
\item{id}{string}

\item{name}{string}

\item{data}{data}

\item{visible}{boolean}

\item{pickable}{boolean}

\item{opacity}{number}

\item{position_format}{unknown}

\item{color_format}{unknown}

\item{auto_highlight}{boolean}

\item{highlight_color}{accessor}

\item{billboard}{boolean}

\item{size_scale}{number}

\item{size_units}{unknown}

\item{size_min_pixels}{number}

\item{size_max_pixels}{number}

\item{background_color}{color}

\item{font_family}{unknown}

\item{font_weight}{unknown}

\item{line_height}{number}

\item{font_settings}{object}

\item{word_break}{unknown}

\item{max_width}{number}

\item{get_text}{accessor}

\item{get_position}{accessor}

\item{get_color}{accessor}

\item{get_size}{accessor}

\item{get_angle}{accessor}

\item{get_text_anchor}{accessor}

\item{get_alignment_baseline}{accessor}

\item{get_pixel_offset}{accessor}

\item{tooltip}{accessor}
}
\description{
Text Layer
}
\seealso{
\url{https://github.com/visgl/deck.gl/blob/v8.3.11/docs/api-reference/layers/text-layer.md}

Other layers: 
\code{\link{arc_layer}},
\code{\link{bitmap_layer}},
\code{\link{column_layer}},
\code{\link{contour_layer}},
\code{\link{cpu_grid_layer}},
\code{\link{geojson_layer}},
\code{\link{gpu_grid_layer}},
\code{\link{great_circle_layer}},
\code{\link{grid_cell_layer}},
\code{\link{grid_layer}},
\code{\link{h3_cluster_layer}},
\code{\link{h3_hexagon_layer}},
\code{\link{heatmap_layer}},
\code{\link{hexagon_layer}},
\code{\link{icon_layer}},
\code{\link{line_layer}},
\code{\link{mvt_layer}},
\code{\link{path_layer}},
\code{\link{point_cloud_layer}},
\code{\link{polygon_layer}},
\code{\link{s2_layer}},
\code{\link{scatterplot_layer}},
\code{\link{screen_grid_layer}},
\code{\link{simple_mesh_layer}},
\code{\link{solid_polygon_layer}},
\code{\link{terrain_layer}},
\code{\link{tile_3d_layer}},
\code{\link{tile_layer}},
\code{\link{trips_layer}}
}
\concept{layers}
