% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_layer.R
\name{tile_layer}
\alias{tile_layer}
\title{\href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/tile-layer.md}{TileLayer} deck.gl layer.}
\usage{
tile_layer(
  id = NULL,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = c(0, 0, 128, 128),
  get_tile_data = NULL,
  max_zoom = NULL,
  min_zoom = 0,
  max_cache_size = NULL,
  ...
)
}
\arguments{
\item{id}{\code{\link{character}}
The id of the layer. Layer ids must be unique per layer \code{type} for deck.gl
to properly distinguish between them.}

\item{data}{\code{\link{data.frame}} | \code{\link[sf:sf]{sf::sf}}}

\item{visible}{\code{\link{logical}}}

\item{pickable}{\code{\link{logical}}}

\item{opacity}{\code{\link{numeric}}}

\item{position_format}{\code{XY} | \code{XYZ}}

\item{color_format}{\code{RGB} | \code{RGBA}}

\item{auto_highlight}{\code{\link{logical}}}

\item{highlight_color}{\code{\link{integer}}}

\item{get_tile_data}{\code{\link[htmlwidgets:JS]{htmlwidgets::JS}}}

\item{max_zoom}{\code{\link{numeric}}}

\item{min_zoom}{\code{\link{numeric}}}

\item{max_cache_size}{\code{\link{numeric}}}

\item{...}{additional layer parameters to pass to deck.gl.
\code{snake_case} parameters will be converted to \code{camelCase}.}
}
\value{
\code{TileLayer} & \code{\link{layer}}
A \href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/tile-layer.md}{TileLayer} layer.
Add to an \link{rdeck} map via \code{\link{add_layer}} or \code{\link{rdeck}}.
}
\description{
\href{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/tile-layer.md}{TileLayer} deck.gl layer.
}
\seealso{
\url{https://github.com/uber/deck.gl/blob/v8.0.16/docs/layers/tile-layer.md}
}
