% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deckgl-layers-doc.R, R/deckgl-layers.R
\name{tile_layer}
\alias{tile_layer}
\alias{add_tile_layer}
\title{Tile Layer}
\usage{
add_tile_layer(
  rdeck,
  ...,
  id = NULL,
  name = "TileLayer",
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  get_tile_data = NULL,
  extent = NULL,
  tile_size = 512,
  max_zoom = NULL,
  min_zoom = 0,
  max_cache_size = NULL,
  max_cache_byte_size = NULL,
  refinement_strategy = "best-available",
  z_range = NULL,
  max_requests = 6,
  tooltip = FALSE
)
}
\arguments{
\item{rdeck}{<\code{\link{rdeck}} | \code{\link{rdeck_proxy}}> An rdeck map instance.}

\item{...}{Additional parameters that will be forwarded to deck.gl javascript without
validation nor processing. All dots must be named and will be \code{camelCased} when serialised.
Layers will raise a warning when dots are used, warning class \code{rdeck_dots_nonempty}.}

\item{id}{<\code{string}> The layer's identifier must be unique for among all layers of the same
type for a map. Defaults to \code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}}, but should be explicitly defined for
updatable layers in a shiny application.}

\item{name}{<\code{string}> Identifies the layer on tooltips and legends. It does
not need to be unique, but should be brief. Defaults to the deck.gl class name for the layer.}

\item{data}{<\code{data.frame} | \code{sf} | \code{string}> The layer's data. Data frames will contain all
columns that are referenced by the layer's accessors. Strings will be interpreted as a URL and
data will be retrieved dynamically in the browser.}

\item{visible}{<\code{logical}> Determines whether the layer is visible or not; also determines
whether any legend elements for the layer will be displayed.}

\item{pickable}{<\code{logical}> Determines if the layer responds to pointer / touch events.}

\item{opacity}{<\code{numeric}> Determines the layer's opacity.}

\item{position_format}{<\code{"XY"} | \code{"XYZ"}> Determines whether each coordinate has two (XY)
or three (XYZ) elements.}

\item{color_format}{<\code{"RGB"} | \code{"RGBA"}> Determines whether the alpha channel of the colours
will be ignored by accessors, making all colours opaque.}

\item{auto_highlight}{<\code{logical}> When \code{TRUE}, the current object \emph{hovered} by the cursor is
highlighted by \code{highlight_color}.}

\item{highlight_color}{<\code{\link{accessor}} | \code{\link{scale}} | \code{\link{color}}> When \code{auto_highlight} and
\code{pickable} are enabled, \code{highlight_color} determines the colour of the currently
highlighted object. If a single colour value is supplied, that colour will be used to
highlight all objects in the layer. Per-object highlighting is achieved with a colour scale,
or a \href{https://dplyr.tidyverse.org/articles/programming.html}{tidy-eval} column of colours.}

\item{get_tile_data}{<\code{\link[htmlwidgets:JS]{JS}}> retrieves the data of each tile.
See deck.gl \href{https://deck.gl/docs/api-reference/geo-layers/tile-layer#gettiledata}{TileLayer}.}

\item{extent}{<\code{c(min_x, min_y, max_x, max_y)}> Tiles in this bounding box
will be rendered at \code{min_zoom}, when zoomed out below \code{min_zoom}.}

\item{tile_size}{<\code{number}> A power of 2 that is the pixel dimensions of the tile.}

\item{max_zoom}{<\code{number}> Tiles above this zoom level are not shown. Defaults to \code{NULL}.}

\item{min_zoom}{<\code{number}> Tiles below this zoom level are not shown. Defaults to \code{0}.}

\item{max_cache_size}{<\code{number}> Maximum number of tiles that can be cached. Defaults to
5x the number of tiles in current viewport.}

\item{max_cache_byte_size}{<\code{number}> Maximum memory used for caching tiles.}

\item{refinement_strategy}{<\code{"best-available"} | \code{"no-overlap"} | \code{"never"}> How the tile
layer refines visibility of tiles. Defaults to \code{"best-available"}.}

\item{z_range}{<\code{c(min, max)}> Array representing the range of heights in the tile.}

\item{max_requests}{<\code{number}> Maximum number of concurrent HTTP
requests across all specified tile provider domains. If a negative number is
supplied no throttling occurs (HTTP/2 only).}

\item{tooltip}{<\code{\link{tooltip}}> Defines the columns (and their order) that will be displayed in
the layer tooltip, if \code{pickable = TRUE}. Supports <\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}>
if a \code{data} is a \code{data.frame}. \code{sfc} columns are always removed.}
}
\description{
Tile Layer
}
\seealso{
\url{https://github.com/visgl/deck.gl/blob/v8.3.11/docs/api-reference/geo-layers/tile-layer.md}

Other geo-layers: 
\code{\link{great_circle_layer}},
\code{\link{h3_cluster_layer}},
\code{\link{h3_hexagon_layer}},
\code{\link{mvt_layer}},
\code{\link{s2_layer}},
\code{\link{terrain_layer}},
\code{\link{tile_3d_layer}},
\code{\link{trips_layer}}

Other layers: 
\code{\link{arc_layer}},
\code{\link{bitmap_layer}},
\code{\link{column_layer}},
\code{\link{contour_layer}},
\code{\link{cpu_grid_layer}},
\code{\link{geojson_layer}},
\code{\link{gpu_grid_layer}},
\code{\link{great_circle_layer}},
\code{\link{grid_cell_layer}},
\code{\link{grid_layer}},
\code{\link{h3_cluster_layer}},
\code{\link{h3_hexagon_layer}},
\code{\link{heatmap_layer}},
\code{\link{hexagon_layer}},
\code{\link{icon_layer}},
\code{\link{line_layer}},
\code{\link{mvt_layer}},
\code{\link{path_layer}},
\code{\link{point_cloud_layer}},
\code{\link{polygon_layer}},
\code{\link{s2_layer}},
\code{\link{scatterplot_layer}},
\code{\link{scenegraph_layer}},
\code{\link{screen_grid_layer}},
\code{\link{simple_mesh_layer}},
\code{\link{solid_polygon_layer}},
\code{\link{terrain_layer}},
\code{\link{text_layer}},
\code{\link{tile_3d_layer}},
\code{\link{trips_layer}}
}
\concept{geo-layers}
\concept{layers}
