% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tooltip}
\alias{tooltip}
\title{Tooltip}
\usage{
tooltip(...)
}
\arguments{
\item{...}{names}
}
\value{
\code{character}
\code{...} converted to a character vector representing names to display.
}
\description{
Tooltip
}
