% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-proptypes.R
\name{tooltip}
\alias{tooltip}
\title{Prop Type: Tooltip}
\description{
If \code{pickable == TRUE}, tooltip defines which columns (and their order) will be displayed
when an object is \emph{hovered}. The tooltip will be displayed as a transposed table (1 row
per column) and is styled according to the \code{\link[=rdeck]{rdeck()}} \code{theme}. The tooltip layout is inspired
by \href{https://kepler.gl}{kepler.gl}.

\if{html}{\out{
  <div class="text-center">
    <img src="https://user-images.githubusercontent.com/391385/103058202-57f26080-45ed-11eb-9131-f09585c1b894.png" />
  </div>
}}

Tooltips support different arguments depending on the value of \code{data}, but the following
arguments are always supported:
\itemize{
\item \code{NULL} | \code{NA} | \code{FALSE} -> no tooltip
\item \code{TRUE} -> all columns (except \code{sfc} columns)
}
}
\section{Data inherits \code{data.frame}}{
When \code{data} is a \code{data.frame}, tooltip supports <\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> (renaming
isn't supported and will fail silently). Expressions such as: \code{c(foo, bar)}, \code{everything()}, \code{matches("foo")},
\code{-unwanted}, \code{where(is.character) & matches("foo")}, \code{1:6} etc. are supported.

All columns referenced in tooltip must exist. Use \code{\link[tidyselect:all_of]{tidyselect::any_of()}} to reference columns that may not
be present in data.
}

\section{Data is a \code{string} or \code{NULL}}{
For cases where data is absent (tooltip wouldn't be displayed anyway), or dynamically loaded
in the client, tooltips may take:
\itemize{
\item a character vector of column names (e.g. c("foo", "bar"))
\item a bare name (e.g. \code{my_col})
\item or a \code{c()} expression of bare names (e.g. \code{c(foo, bar)}).
}

No validation on these columns is possible in this case.
}

\keyword{internal}
