% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_state.R
\name{view_state}
\alias{view_state}
\title{Create a view_state}
\usage{
view_state(center = NULL, zoom = NULL, pitch = NULL, bearing = NULL)
}
\arguments{
\item{center}{\code{\link{numeric}} | \code{\link[sf:st_point]{sf::st_point}}
Center of the viewport on map}

\item{zoom}{\code{\link{numeric}}
Scale = \code{Math.pow(2, zoom)} on map}

\item{pitch}{\code{\link{numeric}}
Camera angle in degrees (0 is straight down)}

\item{bearing}{\code{\link{numeric}}
Map rotation in degrees (0 means north is up)}
}
\description{
Create a view_state
}
\seealso{
\url{https://github.com/uber-archive/viewport-mercator-project/blob/master/docs/api-reference/web-mercator-viewport.md}
}
