% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{visit}
\alias{visit}
\title{Simple expression visitor for translating R into accessors}
\usage{
visit(expr, names, columnar = TRUE)
}
\arguments{
\item{expr}{\code{{name | call | value}}
The expression to visit}

\item{names}{\code{{character}}
Vector of names to expand into either \code{data.frame[expr][index]}
or \code{data[expr]}, depending on the value of \code{columnar} parameter}

\item{columnar}{\code{{logical}}
Is the data passed to layers a columnar table, or an array of objects?}
}
\value{
\code{{JS | name | call | value}}
Either a \link[htmlwidgets:JS]{htmlwidgets::JS} or the original expression
}
\description{
Simple expression visitor for translating R into accessors
}
